improve code. 1. Added a single layout for auth screens. 2. Remove un-necessary svg components. -- we can use them directly using img tags
