+++ .gitignore
@@ -21,7 +21,5 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
-<<<<<<< HEAD
-=======
 
->>>>>>> 9ddff4fbe1c013017aa915e205c479acbf8fe719
+package-lock.json
