--- src/iconComponents/OfcWork.tsx
@@ -1,265 +1,0 @@
-import * as React from 'react';
-
-function SvgOfcWork(props: React.SVGProps<SVGSVGElement>) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 624.04 559.08"
-      {...props}
-    >
-      <defs>
-        <filter
-          id="ofcWork_svg__a"
-          width={1.085}
-          y={-0.236}
-          x={-0.042}
-          height={1.471}
-          colorInterpolationFilters="sRGB"
-        >
-          <feGaussianBlur stdDeviation={15.328} />
-        </filter>
-      </defs>
-      <g fillOpacity={0.996}>
-        <path
-          opacity={0.8}
-          filter="url(#ofcWork_svg__a)"
-          d="M101.31 916.93l588.13-155.97 289.66 66.55-602.3 181.59-275.49-92.17z"
-          transform="translate(-36.835 -69.51) scale(.5709)"
-        />
-        <path d="M42.058 234.757l393.264-42.247v109.733L42.058 387.878V234.757z" />
-        <path
-          fill="#ccc"
-          d="M364.33 215.786l6.564-.911v158.601l-6.563.912V215.786z"
-        />
-        <path
-          fill="#666"
-          d="M358.416 214.792l6.768.99v158.602l-6.768-.99V214.792z"
-        />
-        <path
-          fill="#ccc"
-          d="M42.743 234.237l6.563-1.281v223.01l-6.563 1.282v-223.01z"
-        />
-        <path
-          fill="#666"
-          d="M36.828 232.844l6.768 1.392v223.011l-6.768-1.392v-223.01z"
-        />
-        <path
-          fill="#ccc"
-          d="M503.796 203.746l6.563-1.155v200.997l-6.563 1.154V203.746z"
-        />
-        <path
-          fill="#666"
-          d="M497.881 202.49l6.768 1.254v200.997l-6.768-1.255V202.49z"
-        />
-        <path
-          fill="#c87137"
-          d="M436.521 218.36l67.772-13.305v142.246l-67.806 21.551.036-150.495z"
-        />
-        <path
-          fill="#784421"
-          d="M440.626 277.375l62.028-14.844v79.755l-62.028 17.127v-82.038z"
-        />
-        <path
-          fill="#28170b"
-          d="M261.78 186.927l174.741 32.818v149.102L261.78 325.459V186.924z"
-        />
-        <path
-          fill="#ccc"
-          d="M180.564 266.3l6.563-1.348v234.423l-6.563 1.347V266.299z"
-        />
-        <path
-          fill="#666"
-          d="M174.65 264.838l6.767 1.463v234.423l-6.768-1.463V264.838z"
-        />
-        <path
-          fill="#784421"
-          d="M42.326 238.114L174.82 267.8v162.512L43.468 389.208l-1.142-151.094z"
-        />
-        <path
-          fill="#28170b"
-          d="M21.003 224.464l159.566 36.51.01 9.332-159.573-36.818v-9.024z"
-        />
-        <path
-          fill="#c87137"
-          d="M187.13 266.944l95.162-18.269V410.36l-95.163 27.403V266.944z"
-        />
-        <path
-          fill="#784421"
-          d="M180.564 259.825l334.673-62.456v7.062l-334.673 65.882v-10.488zM440.626 224.875l61.954-12.56V253.1l-61.954 14.843v-43.07z"
-        />
-        <path
-          fill="#502d16"
-          d="M21.003 224.464l333.48-52.506 160.8 25.433-334.72 63.684-159.56-36.611z"
-        />
-        <path
-          fill="#784421"
-          d="M192.872 331.753l82.039-17.127v89.585l-82.039 23.978v-96.436zM194.922 277.146l80.805-14.843v43.07L194.922 322.5v-45.354z"
-        />
-        <path
-          fill="#e6e6e6"
-          d="M223.227 346.202l31.995-5.709v4.512l-31.995 5.71v-4.513z"
-        />
-        <path
-          fill="#ccc"
-          d="M223.227 296.98l31.995-5.71v4.512l-31.995 5.71v-4.513zM464.41 245.587l24.002-4.567v4.512l-24.003 4.567v-4.512zM463.382 288.861l24.002-4.567v4.512l-24.002 4.567v-4.512z"
-        />
-      </g>
-      <g
-        fillRule="evenodd"
-        stroke="#000"
-        strokeMiterlimit={10}
-        strokeWidth={0.036}
-      >
-        <path
-          d="M365.44 127.36c-2.973 2.545-6.604 5.49-10.892 8.948-4.345 3.489-9.29 7.233-14.95 11.664-3.202.571-6.29 1.629-9.72 3.287-3.545 1.572-7.518 3.573-10.491 6.433-2.859 2.8-4.517 5.23-6.632 10.49-2.173 5.289-4.831 12.55-6.204 20.783-1.486 8.118-2.344 19.696-2.344 27.786-.028 8.004 1.058 14.464 2.144 19.839.915 5.202.2 4.345 3.688 11.634 3.43 7.347 9.033 17.753 16.923 31.874 1.086.029 2.116-.056 3.087-.371.972-.372 1.716-.858 2.545-1.572.314 2.287.4 4.403.371 6.232-.2 1.83-.829 3.345-.772 4.86 0 1.4.687 2.916 1.173 3.888.4.914.943 1.315 1.743 1.743.8.343 1.801.572 3.116.572v5.833c-1.515.027-3.173.084-4.66.4-1.514.285-3.172.6-4.659 1.172-.943 12.149-1.4 25.099-1.372 39.735.028 14.58.543 29.96 1.544 47.025 3.573.057 7.203.057 10.891 0 3.688-.143 7.347-.343 11.092-.6l.972-22.154c.628 4.458 1.4 8.89 2.315 13.407.943 4.516 2.201 10.52 3.116 13.607.829 2.973.972 3.287 1.744 4.46.686 1.086 1.429 1.8 2.344 2.143 1.258 7.403 2.172 14.665 3.087 22.012.772 7.375 1.572 15.294 1.944 21.754.257 6.318.029 11.521.2 16.152.143 4.517.4 8.205.772 11.063 1.772.029 3.373.029 4.688 0 1.344-.085 2.401-.115 3.288-.2v1.372c-3.03 1.087-6.232 3.26-9.72 6.404-3.545 3.23-8.519 8.261-11.063 12.435-2.658 4.058-4.088 8.433-4.288 11.463-.2 2.916.944 5.632 3.316 5.832 2.373.057 6.604-2.4 10.492-5.06 3.802-2.772 7.804-6.317 12.235-11.063h6.031c1.401-1.515 2.859-3.03 4.26-4.66 1.315-1.658 3.344-3.23 4.088-5.06.628-2.029.686-4.316.4-6.402-.258-2.058-.8-4.002-1.744-5.832 1.2-.714 2.001-1.257 2.516-1.944.457-.714.628-1.257.4-1.943.486-.115 1.086-.115 2.115 0 .944.028 2.173.2 3.688.4-1.2 1.058-2.116 2.344-2.516 3.859-.429 1.486-.314 3.173-.2 5.06 0 1.715.172 3.573.4 5.46.773 2.03 1.973 3.773 3.489 5.231 1.486 1.487 3.401 2.573 5.631 3.488l6.032-1.343c2.144 2.716 5.175 5.203 9.32 7.575 4.087 2.344 11.72 6.174 15.179 6.403 3.373.086 5.603-2.344 5.03-5.632-.771-3.401-5.46-9.862-9.118-13.978-3.774-4.231-9.148-8.233-12.721-10.577-3.517-2.287-6.375-3.402-8.462-3.23l.572-2.716c1.43-.258 2.83-.486 4.088-.572 1.229-.172 2.43-.2 3.487-.2-1.029-12.635-1.258-24.27-.571-35.562s2.058-21.697 4.459-31.674c1.086-.143 1.915-1.315 2.916-3.744.829-2.431 1.343-2.431 2.544-10.693 1.143-8.518 2.8-27.042 4.66-39.249 1.743-12.235 3.916-23.011 6.203-33.217 3.345-.572 5.774-1.658 7.404-3.374 1.543-1.771 1.515-4.002 2.115-6.802.572-2.86 1.03-6.146 1.372-10.091 1.973-.144 3.345-.601 4.46-1.373.972-.915 1.458-1.744 1.772-3.488.285-1.802-.286-4.802.172-6.803.429-2.001 1.172-3.688 2.344-5.06.685.772 1.658 1.486 2.916 2.144 1.2.629 2.687 1.115 4.46 1.571 4.83-19.553 8.575-34.647 11.462-46.396 2.83-11.663 4.831-16.065 5.46-23.126.572-7.118.029-11.435-1.944-19.239-2.115-7.947-7.146-20.954-10.12-27.557-3-6.575-3.258-8.776-7.374-11.663-4.23-2.945-9.92-4.889-17.495-5.832-3.06-2.945-6.975-5.546-11.863-8.147-4.945-2.602-10.663-5.004-17.466-7.204-16.954-.689-33.933-1.375-50.914-2.09z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          fill="#fff"
-          d="M353.291 139.057l13.464-10.662 49.426 1.658 11.464 5.002c-2.83 2.773-5.832 6.118-9.319 10.148-3.516 3.974-8.233 9.863-11.292 13.808-3.03 3.83-5.345 6.975-6.832 9.319-1-4.946-2.172-9.176-3.83-13.15-1.63-4.002-4.46-8.06-5.832-10.319-1.43-2.287-2.315-3.43-2.658-3.317-2.144 3.402-4.231 7.32-6.147 11.635-2.03 4.374-3.716 9.005-5.488 14.322a186.101 186.101 0 00-8.319-12.635c-2.887-4.06-6.46-8.348-8.833-11.006-2.517-2.659-4.346-4.26-5.804-4.803z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          fill="#a80000"
-          d="M387.562 143.544c-2.315 5.288-4.059 9.52-5.345 12.807-1.315 3.287-2.144 5.488-2.487 6.832 1.029 2.515 2.03 4.63 3.001 6.66.972 1.944 1.916 3.66 2.83 5.146.343.114.915.114 1.83 0 .886-.114 3.001-.886 3.487-.657.315.085-.228 1.057-1.172 1.486-1 .343-2.515.658-4.66.829-1.4 5.231-2.658 10.148-3.658 14.98-1.058 4.83-1.887 9.233-2.487 13.635.8 1.258 2.058 3.402 4.002 6.66a929.002 929.002 0 017.318 12.493c.715-2.773 1.515-5.717 2.83-9.148a162.577 162.577 0 014.488-11.148c-1.344-5.403-2.46-10.635-3.345-15.809-.886-5.231-1.543-10.205-1.972-15.15 1.658-1.859 2.944-3.374 3.973-4.66 1.001-1.372 1.63-2.23 2.002-3.002-1.001-4.545-2.344-8.69-4.146-12.32-1.83-3.688-3.916-6.861-6.489-9.634z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          fill="#fff"
-          d="M353.119 139.709l-1.515 1c3.63 8.948 6.689 16.78 9.49 23.956 2.688 7.089 3.974 11.92 6.833 18.466 2.801 6.49 6.089 13.179 9.976 20.297.886-5.946 1.83-11.263 2.83-15.951.944-4.774 1.944-8.777 3.002-12.321-1.43-3.173-2.344-5.575-3.173-7.318-.8-1.773-1.258-2.83-1.487-3.173-.714 2.144-1.286 3.458-1.83 4.173-.514.686-.685.772-1.514-.171-.915-1.03-2.144-3.03-4.145-6.004-2.202-3.086-5.232-8.318-8.319-12.148-3.087-3.831-6.432-7.433-10.148-10.806zM428.33 135.706l1.143.343c.714 1.86.658 4.489-.486 8.32-1.172 3.802-3.516 8.89-6.146 14.15-2.744 5.174-6.946 11.491-9.834 16.637-2.944 4.974-5.344 8.947-7.66 13.464-2.344 4.574-4.546 8.89-6.489 13.493-1.373-6.003-2.516-11.12-3.317-15.809-.857-4.688-1.372-8.604-1.658-11.977.8-1.144 1.63-2.259 2.487-3.345.801-1.172 1.63-2.172 2.488-3.316.171.543.4.943.685 1.315.229.343.486.83 1.002.686.428-.257.428-.171 2-2.172 1.516-2.144 4.66-6.66 7.462-10.149 2.772-3.573 6.602-7.718 9.176-10.662 2.43-2.945 4.287-5.232 5.803-6.975 1.572-1.83 2.543-3.202 3.344-4.003z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          fill="#294b70"
-          d="M350.461 141.543a87.477 87.477 0 00-5.489 3.83c-1.743 1.315-3.316 2.487-4.83 3.831 1.543 3.773 2.887 7.318 4.144 10.634 1.172 3.345 1.916 6.175 3.345 9.148 1.4 2.887 3.116 5.574 5.146 8.176.486.571 1.372.914 2.658 1.143 1.258.143 2.887.143 5.003 0 .486 1.03.8 2.544.829 4.66.028 2.086-.143 4.66-.515 7.832.543 1.287 2.03 3.802 4.345 7.662 2.344 3.887 6.661 11.005 9.634 15.293a87.677 87.677 0 007.661 9.663c2.316 2.6 4.202 4.345 6.003 5.66l1.487-3.831c-3.774-4.917-7.233-10.091-10.491-15.637-3.316-5.603-5.946-10.863-8.977-17.638-3.173-6.918-6.317-15.35-9.319-22.983-3.087-7.718-6.86-18.01-8.661-22.612-1.802-4.632-2.46-6.203-1.973-4.831zM430.99 136.707c2.458.886 4.802 2.201 7.147 3.83 2.344 1.687 4.545 3.545 6.832 6.004-1.258 3.43-2.773 7.032-4.66 11.15-1.943 4.086-4.831 10.147-6.66 13.12-1.887 2.773-2.373 3.001-4.317 4.002-2.002 1-4.46 1.63-7.49 2-.63 1.058-.8 2.774-.514 5.146.285 2.344.972 5.346 2.172 9.005-2.745 4.174-5.632 8.29-8.661 12.464-3.06 4.145-5.576 7.89-9.635 12.32-4.202 4.431-9.176 8.948-15.15 13.98-.03-.887.6-3.23 2-7.147 1.458-4.031 3.974-11.378 6.318-16.323 2.23-4.917 4.03-8.09 6.832-13.15 2.66-5.26 6.433-12.32 9.462-17.638 3.03-5.317 5.975-9.805 8.49-14.122 2.459-4.402 4.574-8.289 6.004-11.977 1.286-3.774 1.658-8.033 2.001-10.148.258-2.144.2-2.946-.171-2.516z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          d="M333.88 89.372c-.057 3.116.457 6.546 1.43 10.406.943 3.802 2.344 8.49 4.316 12.635 1.944 4.087 4.145 8.29 7.032 12.064 2.773 3.83 5.946 7.174 9.634 10.347 3.63 3.002 7.604 5.775 12.378 7.776 4.802 2.002 10.834 3.488 16.094 4.03 5.174.372 10.034-.057 14.808-1.286 4.746-1.315 9.405-3.287 13.636-5.917 4.145-2.716 7.947-6.718 11.092-9.92 2.916-3.287 5.003-6.145 7.031-9.461a76.982 76.982 0 005.317-10.349c1.43-3.573 2.545-7.575 3.459-10.805.857-3.26 1.429-6.09 1.715-8.605.829.4 1.887.6 3.173.6 1.286-.028 3.059.03 4.43-.713 1.316-.858 2.46-2.059 3.317-4.031.744-2.002 1.315-5.146 1.286-7.89-.114-2.916-.943-6.661-1.715-8.919-.858-2.23-1.83-3.602-3.03-4.46-1.314-.914-2.887-1.028-4.287-1-1.401-.028-2.659.172-3.888.714.743-2.83.743-6.26 0-10.634-.83-4.374-2.26-10.091-4.603-15.38-2.487-5.374-7.118-12.663-9.49-16.465-2.373-3.745-2.401-4.374-4.402-6.204-2.002-1.944-4.288-3.916-7.52-5.26-3.316-1.515-7.346-2.572-11.977-3.173-4.716-.714-10.12-.857-15.78-.314-5.803.514-12.55 1.715-18.238 4.03-5.746 2.344-10.834 5.832-15.38 9.92-4.63 4.03-8.518 8.976-11.663 13.95-3.173 4.83-5.403 11.291-7.032 15.093-1.744 3.716-2.23 5.203-2.887 7.318-.658 2.144-.944 3.717-.858 5.031.086 1.173.372 1.802 1.172 2.002-1.972-.143-3.516.285-4.745 1.172-1.258.858-1.973 2.316-2.602 4.146-.743 1.801-1.486 3.974-1.572 6.632-.057 2.63.029 6.832 1.143 9.062 1.03 2.087 3.602 3.345 5.175 4.002 1.573.571 2.802.628 4.031-.144z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          fill="#e8cdba"
-          d="M337.311 64.643c1.83-.972 3.802-2.259 5.918-3.859 2.058-1.686 4.63-3.802 6.603-5.888 1.944-2.288 3.488-4.804 4.745-6.775 1.115-1.887 1.744-3.46 2.002-4.603 4.602 1.172 8.89 1.83 13.092 2.316 4.202.285 8.176.256 11.92.143 3.631-.115 6.376.028 9.777-.572 3.374-.658 7.033-2.087 10.206-3.173 3.088-1.173 5.66-2.116 8.061-3.602 2.287-1.572 4.117-3.316 5.604-5.46 1.172 2.458 2.372 5.06 3.716 7.918 1.344 2.802 2.63 5.86 4.174 9.062 1.486 3.173 3.06 7.118 4.602 9.948 1.402 2.745 2.887 5.06 4.03 6.603 1.087 1.43 1.487 1.602 2.43 2.173.915.486 1.773.857 2.889 1-.2.515-.258 1.086-.286 1.715-.057.515.028 1.802.143 2.002.085.143.228-.086.429-.572.143-.486.228-1.458.572-2.144a8.435 8.435 0 011.286-2.315c.486-.83 1.201-1.744 2.03-2.288.772-.6 1.772-1.085 3.002-1.142 1.172-.172 3.087-.315 4.316.571 1.2.886 2.316 2.888 3.03 4.889.572 1.886.858 4.087.858 6.602.028 2.46-.257 6.033-.858 8.062-.686 1.858-1.83 3.06-3.316 3.716-1.601.657-4.403.801-5.746.314-1.343-.571-1.972-2.858-2.43-3.316-.6-.457-.772-.115-.743.572-.029.572.972 1.372 1.029 3.173.028 1.744-.086 3.83-.886 7.461-.972 3.574-2.488 9.435-4.431 13.837-2.001 4.345-4.403 8.29-7.09 12.063-2.658 3.688-5.46 7.061-8.947 10.234-3.63 3.116-8.348 6.432-12.493 8.52-4.23 1.943-7.775 3.086-12.492 3.544-4.86.457-10.977.2-15.923-.772-5.03-1.086-9.205-2.916-13.607-5.489-4.43-2.687-9.005-6.346-12.635-10.206-3.688-3.973-6.718-8.976-9.09-13.035-2.402-4.145-3.888-7.86-5.118-11.377-1.343-3.63-2.086-7.004-2.544-9.662-.486-2.63-.571-4.717-.257-6.146.858-.973 1.515-1.802 1.858-2.602.343-.857.286-2.2.143-2.171-.286-.057-.8 1.4-1.43 2.171-.657.745-1.429 2.03-2.43 2.431-1.228.257-2.8.257-4.173-.429-1.4-.744-2.858-2.002-3.602-3.887-.771-1.887-.972-4.946-.714-7.319.171-2.515.972-5.431 1.858-7.203.829-1.83 2-2.63 3.173-3.288 1.086-.686 2.458-.514 3.459-.572.829-.115 1.315-.228 2 0 .801.115 1.602.343 2.288 1.001.543.543.743 2.516 1.029 2.716.2.114.4-.772.286-1.573-.203-.86-.603-1.889-1.317-3.317z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          d="M329.706 75.734c.2-.886.43-2.259 1.001-3.03.543-.8 1.258-1.686 2.03-2.002.6-.428 1.372-.943 2-.428.544.486 1.63 2.459 1.716 3.144.057.515-1 .573-1.287.286-.4-.429-.543-1.801-.857-2.144-.372-.4-.886-.429-1.43.144-.571.513-1.886 2-2.172 3.144-.4 1.058-.372 2.544-.143 3.458.114.858.486 1.43 1 1.716.487.142 1.716-.658 2.03-.43.229.23 0 1.316-.572 1.602-.628.143-2.515.115-3.03-.43-.629-.628-.4-2.029-.428-2.886-.143-.829-.058-1.4.142-2.144zM447.538 78.364c-.115-.943-.228-2.287-.571-3.144-.458-1-1.058-1.802-1.687-2.287-.573-.457-1.286-1.173-1.915-.715-.744.343-2.059 2.287-2.173 2.86-.2.543.8.743 1.23.485.371-.343.685-1.743 1.172-2.001.372-.372.914-.257 1.4.343.487.629 1.402 2.258 1.659 3.459.171 1.114-.085 2.515-.343 3.43-.315.743-.773 1.4-1.258 1.543-.572 0-1.686-.829-1.915-.715-.286.144-.286 1.315.314 1.658.544.201 2.373.458 3.06 0 .572-.543.714-1.943.857-2.772.17-.772.228-1.372.17-2.144z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          d="M444.108 75.591c-.114.286.657 1.458.886 2.287.114.8.172 2.173.286 2.46.143.143.485-.63.572-1.315 0-.744-.343-2.316-.572-2.86-.314-.63-1.143-.858-1.172-.572zM377.417 68.816c0 1.916 1.544 3.46 3.46 3.46s3.458-1.544 3.458-3.46c0-1.886-1.543-3.43-3.459-3.43s-3.459 1.544-3.459 3.43zM391.366 68.96a3.452 3.452 0 003.459 3.458c1.887 0 3.43-1.543 3.43-3.459a3.44 3.44 0 00-3.43-3.43c-1.915 0-3.46 1.544-3.46 3.43zM373.415 59.754c1-.43 3.488-.914 5.46-1.143 1.801-.314 5.146-.372 5.746-.286.543.086-.257.372-2.172.858-2.03.4-7.833 1.486-9.348 1.601-1.487.027-.715-.715.314-1.03zM395.654 59.211c.972.029 3.658.372 5.517.915 1.858.429 4.859 1.514 5.46 1.83.457.257-.429.343-2.344.028-2.03-.4-7.833-1.486-9.262-1.944-1.487-.457-.4-.915.629-.829zM358.322 46.776c.944.43 3.46.83 5.46 1.315 1.916.4 5.174.8 6.032.858.658-.172-.028-.772-1.286-1.144-1.258-.314-4.345-.457-6.06-.743-1.802-.372-3.603-1.315-4.289-1.286-.83-.057-.744.571.143 1zM415.95 41.745c.4 1.315 1.144 3.545 2.002 5.031.658 1.373 2.2 3.345 2.316 3.173-.029-.314-.973-3.116-1.744-4.888-.886-1.858-2.201-5.689-2.716-6.203-.4-.544-.315 1.486.142 2.887zM370.958 48.52c0 .229.543.429 1.229.429.657 0 1.229-.2 1.229-.429 0-.257-.572-.429-1.23-.429-.686 0-1.228.172-1.228.429zM420.897 51.036c-.228.114-.115.686.2 1.257.343.6.772.972.972.858.2-.115.116-.686-.2-1.258-.343-.6-.772-.972-.972-.857zM364.64 103.293l1.286-.572a29.798 29.798 0 002.316 7.46c.972 2.316 2.315 4.546 4.145 6.461 1.858 1.801 4.088 3.288 6.918 4.46 2.801 1.085 6.317 2.516 9.776 2.315 3.345-.314 7.376-2.001 10.207-3.745 2.687-1.743 4.716-4.431 6.316-6.747 1.544-2.515 2.373-5.46 3.145-7.631.686-2.173 1.201-3.945 1.315-5.318l1.715.143c-.858 3.002-1.83 5.918-3.03 8.634-1.172 2.772-2.516 5.46-4.145 7.746-1.687 2.231-3.602 3.86-6.203 5.318-2.716 1.286-6.575 2.487-9.92 2.6-3.402 0-7.204-.6-10.32-2.143-3.23-1.63-6.546-4.46-8.633-7.06-2.115-2.574-3.116-6.262-3.888-8.205-.83-2.03-1.2-3.23-1-3.716z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          d="M363.353 103.15c.8-.486 1.658-.715 2.573-1 .943-.373 2.915-1.43 3.173-1.174.2.2-1.03 1.487-2.03 2.316-1.2.772-3.345 1.744-4.46 2.144-1.114.228-2.344-.057-2.143-.43.142-.456 1.915-1.37 2.887-1.856zM412.607 99.687c-1.029-.2-1.8-.456-2.744-.456-1.115-.315-3.173-.629-3.316-.486-.143.229 1.143 1.258 2.43 1.858 1.143.515 3.659 1.058 4.803 1.172 1.172 0 2.143-.543 2.03-.857-.258-.373-2.202-1.002-3.203-1.23z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          fill="#78605c"
-          d="M355.749 43.117c-.572 1.801-1.544 3.458-2.602 5.288-1.143 1.772-2.344 3.23-4.145 5.26-1.944 1.887-5.002 4.63-7.175 6.375-2.287 1.686-4.402 3.258-5.917 3.83-1.573.486-2.659 0-3.145-.8-.543-.915-.171-2.26.343-4.375a43.93 43.93 0 012.687-8.519c1.515-3.659 3.717-8.947 6.26-13.006 2.43-4.203 5.118-8.005 8.405-11.52 3.202-3.603 6.861-6.862 10.977-9.434 4.06-2.573 8.176-4.375 13.122-5.718 4.83-1.401 11.091-2.258 15.808-2.572 4.602-.372 8.119-.029 12.092.571 3.83.458 7.861 1.058 11.321 2.687 3.488 1.573 6.774 4.517 9.205 7.062 2.23 2.459 3.173 5.031 4.831 7.833 1.743 2.8 3.716 5.974 5.26 8.747 1.43 2.602 2.172 4.002 3.345 7.061 1.057 3.03 2.544 7.72 3.258 10.977.63 3.116.744 5.546.773 7.747-.029 2.116-.144 3.83-.657 5.146-.63 1.258-1.344 2.373-2.46 2.573-1.142.027-2.487-.2-4.03-1.887-1.658-1.802-3.344-4.975-5.174-8.404-1.83-3.574-3.659-7.833-5.603-12.235-2.059-4.403-4.031-8.89-6.26-13.78.913-1.43 1.542-3.144 1.886-5.26.285-2.258.143-6.489.114-7.518-.114-.943-.371.429-.543 1.686-.257 1.23-.086 3.202-.572 5.26-.544 1.973-1.143 4.403-2.572 6.518-1.602 1.944-3.374 3.66-6.518 5.375-3.288 1.658-8.462 3.658-12.55 4.803-4.117 1.086-7.519 1.429-11.749 1.8-4.316.2-9.69.086-13.693-.114-4.03-.172-7.318-.544-9.976-1.115-2.659-.8-4.431-1.83-6.09-2.973-1.829-1.257-4.001-3.458-4.544-3.744-.572-.23.543 1.686 1.143 2.43.515.6 1.258.914 2.115 1.601.857.6 1.83 1.458 3.03 2.344z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          d="M340.74 57.867c.372-.685 2.717-2.058 3.831-3 1.001-.916 1.601-1.487 2.688-2.602 1.086-1.258 3.316-4.145 3.916-4.46.4-.343.028 1.287-.772 2.46-.858 1.172-2.43 2.887-3.83 4.373-1.43 1.43-3.402 3.288-4.374 3.917-1.03.428-1.887-.06-1.458-.688zM338.397 60.098c.086.171.543.057 1-.258.458-.286.744-.657.63-.829-.087-.143-.544-.028-1.001.258-.457.314-.744.685-.63.829zM355.292 37.772c.83.657 3.316 2 5.374 2.915 1.887.744 4.088 1.487 6.604 1.916 2.516.343 7.49.486 8.633.428 1.086-.143-.572-.543-2.002-.771-1.544-.2-4.717-.315-6.746-.573a39.99 39.99 0 01-5.375-1.543 32.094 32.094 0 01-3.802-1.686c-1.057-.544-1.915-1.43-2.344-1.458-.486-.144-1.229.114-.342.772z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          d="M364.696 38.886c.115.458 1 2.059 1.573 2.801.514.715 1.257 1.286 1.457 1.258.086-.143-.571-1.258-.886-1.915-.457-.714-.829-1.944-1.229-2.259-.372-.4-1.085-.4-.915.115zM362.467 39.572c-.058.286.457 1.143.886 1.686.343.43 1.343 1.315 1.458 1.23.085-.172-.686-1.43-.887-1.915-.2-.515-.428-1.03-.571-1.115-.258-.2-.914-.2-.886.114zM388.912 40.773c1.744-.886 7.204-2.516 10.32-3.802a99.004 99.004 0 008.06-3.83c2.173-1.344 4.003-2.688 5.376-3.688 1.372-1.144 2.23-2.002 2.915-2.344.63-.258 1.115-.115.8.543-.428.544-1.886 1.658-3.144 2.687-1.286.915-2.744 2.002-4.602 3.145-1.944 1.114-4.517 2.6-6.947 3.802-2.544 1.114-5.745 2.23-7.947 3.03-2.287.714-4.488 1.458-5.288 1.571-.859.03-1.373-.342.457-1.114zM389.595 22.049c1.458-.315 4.917-.572 8.176-.658 3.2-.142 7.375-.142 10.52.115 3 .172 6.203.801 7.518 1.23 1.23.4.829 1.142.114 1.229-.772 0-2.6-.773-4.716-1.001-2.258-.257-5.575-.485-8.29-.457-2.744 0-5.574.114-7.832.343-2.4.086-4.716.372-5.718.315-.887-.144-1.286-.887.228-1.116zM392.716 11.843c1.543.544 8.376 1.401 12.12 2.688 3.717 1.285 8.29 4.373 10.207 5.174 1.83.686 1.543-.172 1-.686-.629-.658-2.344-1.515-4.374-2.573-2.115-1.114-5.174-2.487-7.86-3.458-2.83-1-6.519-1.83-8.405-2.03-1.889-.23-4.29.285-2.688.885zM419.5 22.506c.486.886 3.344 3.374 4.602 5.032 1.287 1.6 2.144 3.83 2.917 4.602.572.714 1.343.6 1-.229-.428-.914-1.772-3.402-3.03-5.03-1.286-1.688-3.43-4.032-4.345-4.718-1-.8-1.716-.571-1.144.343zM419.841 28.881c.543.458 2.4 3.173 3.488 5.175.973 1.943 1.715 4.66 2.23 6.26.315 1.344.2 2.345.115 2.688-.115.514-.115.629-.458-.115-.371-.914-.972-3.402-1.686-5.26-.914-1.972-2.43-4.83-3.116-6.26-.688-1.573-1.03-2.917-.573-2.488zM430.164 59.383c.4.886 1.887 3.516 2.916 4.802.886 1.115 1.858 1.83 2.572 2.144.657.172 1.143-.428 1.23-.571-.029-.172-.801-.029-1.23-.344-.543-.4-1.143-1.03-1.686-1.772-.629-.857-1.343-2.087-1.802-2.916-.543-.8-.914-1.658-1.229-1.801-.457-.285-1.086-.43-.77.458zM437.446 65.615c.143.143.571-.057.972-.429.4-.372.6-.773.457-.915-.114-.143-.543.057-.943.429s-.6.772-.486.915zM415.037 15.074c.085-1.23.857-3.716 1.114-5.603.172-1.887-.457-4.375-.229-5.375.144-.972.887-1.2 1.23-.686.228.457.457 1.686.457 3.259-.028 1.515-.143 4.087-.456 5.717-.373 1.543-1.002 3.173-1.344 3.688-.372.371-.887.143-.772-1zM417.84 15.96c.715-1.058 3.345-3.173 4.803-4.802 1.401-1.773 2.488-3.489 3.488-5.26.83-1.887 1.458-4.747 2.001-5.489.515-.8 1.515-.057 1.344 1.23-.315 1.314-1.544 4.287-2.802 6.603-1.229 2.2-3.344 4.916-4.687 6.489-1.459 1.486-3.173 2.458-3.83 2.687-.774.085-1.03-.4-.317-1.458zM421.302 18.104c.6-.686 2.087-2.315 3.802-2.917 1.601-.629 5.203-.885 6.174-.771.915.085.286.972-.572 1.343-.971.285-3.487-.2-4.802.429-1.4.514-2.487 2.4-3.259 2.916-.8.429-1.286.028-1.458-.115-.285-.17-.543-.313.115-.885zM368.784 139.685c.772.544 2.773 1.686 4.688 2.43 1.858.686 4.317 1.286 6.375 1.687 2.001.314 4.574.657 5.403.572.686-.143-.057-.858-1.143-1.144-1.315-.343-3.802-.4-5.804-.829-2.058-.572-4.259-1.543-5.831-2.144-1.572-.714-2.973-1.571-3.545-1.686-.629-.087-.972.427-.143 1.114zM386.967 143.544c-.143 0-.458.343-.286.6.057.115.514.543.715.515.2-.057.371-.686.4-.801-.172-.257-.6-.4-.83-.314zM366.669 136.993c-.257.057-.4.371-.315.629-.028.114.543.543.715.514.172-.085.429-.657.429-.83-.172-.256-.6-.4-.83-.313z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          fill="#294b70"
-          d="M338.511 149.976c-3.373.772-6.489 1.915-9.576 3.43-3.145 1.343-6.204 2.887-8.72 5.403-2.543 2.487-4.116 4.517-6.145 9.834-2.144 5.345-4.889 14.665-6.26 22.011-1.402 7.29-1.687 14.637-1.859 21.383-.2 6.69.229 13.321.743 18.467.4 5.06.658 8.061 1.858 11.778 1.115 3.601 3.202 6.632 5.032 10.09 1.743 3.375 3.373 6.261 5.66 10.321 2.201 4.06 4.774 8.576 7.718 13.865 2.03-2.773 4.117-5.145 6.146-7.003 2.03-1.916 3.745-3.23 6.032-4.288 2.173-1.172 4.545-2.03 7.118-2.573-.657-2.6-1.486-6.003-2.83-10.32-1.315-4.403-3.916-11.292-5.031-15.494-1.23-4.231-1.487-7.604-2.087-9.577-.629-1.972-.429-1.8-1.715-2.23-1.515-.543-5.603-.714-6.404-.972-.771-.4.915-1.029 2.087-1.114 1.2-.058 3.688.4 4.803.371.914-.2.772-.057 1.115-1.343.257-1.401.428-4.86.857-6.747.372-1.915.944-3.259 1.715-4.688.715-1.544 2.03-2.973 2.83-4.288.715-1.372 1.43-2.601 1.83-3.802-.743-.629-1.544-1.515-2.087-2.773-.686-1.315-1.63-4.088-1.6-4.688-.03-.6 1.057.429 1.6 1.23.515.828.6 2.458 1.487 3.458.857.858 2.515 1.287 3.43 1.715.686.286.429.343 1.23.486.771.115 3.544-.057 3.801.258.143.257-1.343 1.143-2.458 1.343-1.115.086-2.544 0-4.174-.486a16.762 16.762 0 01-1.944 4.774c-.943 1.344-2.601 2.144-3.316 3.83-.743 1.716-.857 4.46-1 6.261-.258 1.601-.086 2.916-.372 3.945-.371.972-1.372 1.6-1.458 2.316-.114.628.543.171.972 1.715.257 1.458.372 3.802 1.23 7.146.857 3.23 2.83 8.69 4.059 12.407 1.086 3.63 2.03 6.832 2.83 9.49.686 2.602 1.372 4.547 1.715 6.147.314 1.571.314 2.601.114 3.201-.371.8-.057 1.515.63 2.572.685.973 2.572 1.86 3.687 3.202.915 1.23 1.658 2.544 2.573 4.402.914 1.802 2.115 4.631 3.058 6.518.887 1.858 2.144 2.973 2.459 4.431.2 1.373.314 3.002-.858 3.688-1.172.544-3.259.457-6.146-.257 1.23.972 1.772 1.772 1.973 2.83.142 1 .028 2.2-1.23 2.944-1.315.601-3.487.886-6.632.857l.115 5.918c.914-.143 2.03-.143 3.573 0 1.486.143 3.287.429 5.403.857v8.348c4.345-.315 8.118-.514 11.434-.485 3.26-.03 5.746.457 8.119.371 2.258-.171 3.487.486 5.889-1.23 2.43-1.857 5.26-4.83 8.604-9.347-2.144-6.432-3.516-13.292-4.43-20.896-.944-7.661-1.23-16.437-.858-24.441.371-8.062 1.315-15.866 2.944-23.641-3.001-2.544-5.946-5.546-8.833-9.12-2.916-3.601-5.374-7.175-8.605-12.263-3.373-5.174-6.917-11.32-11.063-18.553.315-2.286.515-4.345.6-6.289.03-2 .03-3.659-.114-5.288-1.743.429-3.258.486-4.66.171-1.371-.371-2.6-1-3.687-2.23-1.2-1.343-2.315-3.258-3.459-5.516-1.2-2.373-2.172-4.717-3.687-8.348-1.6-3.746-3.487-8.064-5.775-13.494z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          fill="#294b70"
-          d="M446.625 147.347c2.602.115 5.231.658 8.004 1.516 2.802.743 5.86 1.6 8.29 3.658 2.172 2.03 3.459 4.574 5.403 8.948 2.03 4.345 4.46 11.777 6.518 17.238 1.972 5.431 4.03 11.12 5.174 15.208 1.03 3.945 1.173 5.488 1.202 8.947-.086 3.402-.172 6.06-1.344 11.95-1.258 5.917-4.03 16.494-5.689 23.47-1.715 6.974-2.715 11.348-4.23 17.866-1.515 6.403-3.174 13.092-4.86 20.382-.944-2.544-2.4-4.916-4.487-7.347-2.087-2.458-5.232-5.403-7.891-7.06-2.744-1.773-6.947-1.63-8.405-3.117-1.486-1.543-.943-3.088-.685-5.832.172-2.859.714-7.175 1.63-10.863.743-3.688 1.83-7.433 3.144-11.12 1.343-3.86 4.174-8.719 4.86-11.263.572-2.516-.344-2.602-1.202-3.802-.914-1.315-2.743-2.544-3.945-3.802-1.315-1.315-2.744-2.573-3.516-3.774s-1.029-2.23-1.086-3.516c-.085-1.458.457-4.173.4-4.745-.171-.658-1.142.257-1.372 1.2-.229.887-.028 2.888-.4 4.231-.573 1.23-1.83 1.887-2.573 3.516-.772 1.544-1.658 5.089-1.772 6.118-.085.915.572.6 1.086-.286.4-1 .715-3.973 1.373-5.288.543-1.344 1.343-2.116 2.172-2.43.572 1.429 1.343 2.63 2.43 3.916 1 1.258 2.745 2.287 3.802 3.402.972.886 2.087 1.486 2.459 2.573.286 1.057.343 1.944-.286 4.087-.714 2.144-2.4 5.775-3.516 8.805-1.115 2.916-2.116 5.174-3.002 9.09-.943 3.974-2.115 11.006-2.43 14.265-.314 3.059-.115 3.488.258 4.603.256 1.143.629 1.715 1.229 2.03-.573.886-1.43 1.743-2.46 2.459-1.142.657-2.743.771-3.916 1.886-1.229 1.143-1.715 3.202-3.001 4.746-1.401 1.429-3.917 2.258-4.888 3.802-1.029 1.486-1.173 3.487-1.086 4.888-.029 1.23.343 2.172 1.23 2.716.914.513 2.086.457 3.944.143 0 .428-.4.8-1.086 1.343-.858.4-2.716.8-3.259 1.63-.572.857-.457 2.286 0 3.401.457 1.03 1.343 2.23 2.43 2.716 1.001.372 2.172.029 3.402-.143 1.172-.2 2.315-.6 3.544-1.086.172 1.63.544 2.716 1.344 3.259.772.457 2.144.314 3.116 0 .886-.343 1.658-.972 2.315-1.887.4.372.972.686 1.487.943.572.258 1.086.4 1.772.544-.372 2.687-.657 5.26-1.086 7.604-.429 2.372-.373 4.516-1.373 6.431-1.086 1.773-1.572 3.26-4.888 4.346-3.459.972-9.49 1.601-15.323 1.772-5.889.144-14.236.029-19.41-.685-5.145-.715-8.86-1.973-11.263-3.66-2.173-4.688-3.83-10.119-5.288-16.294-1.487-6.346-2.83-13.464-3.26-20.896-.485-7.461-.085-16.553.544-23.07.514-6.518 1.543-11.807 2.858-15.866 2.659-1.544 5.518-3.63 8.69-6.375a113.357 113.357 0 0010.035-10.177c3.258-3.859 6.173-8.861 9.09-12.806 2.8-3.974 5.174-7.719 7.46-11.12-.943-3.43-1.658-6.004-2.03-8.148-.4-2.115-.572-3.601-.4-4.46 2.488.03 4.66-.514 6.633-1.514 1.886-1.03 3.201-2.058 5.03-4.745 1.802-2.888 3.746-7.862 5.832-11.95 2.03-4.145 4.203-8.204 6.518-12.52z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          d="M390.739 251.03c.257 2.429 1.886 4.23 3.63 4.058 1.744-.2 2.973-2.315 2.716-4.745-.257-2.401-1.887-4.23-3.66-4.03-1.743.172-2.943 2.286-2.686 4.716z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          fill="#294b70"
-          d="M391.707 251.339c.172 1.515 1.201 2.658 2.317 2.544 1.085-.114 1.858-1.458 1.686-2.973-.143-1.544-1.173-2.687-2.287-2.573-1.115.144-1.859 1.459-1.716 3.002z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          d="M390.453 275.47c.257 2.402 1.886 4.231 3.63 4.031 1.773-.171 2.973-2.287 2.716-4.716-.257-2.43-1.887-4.23-3.63-4.06-1.773.2-2.973 2.316-2.716 4.745z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          fill="#294b70"
-          d="M391.453 275.756c.143 1.544 1.172 2.687 2.287 2.573 1.115-.144 1.858-1.458 1.715-3.001-.171-1.516-1.201-2.659-2.316-2.545-1.086.115-1.858 1.458-1.686 2.973z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          fill="#e0e0e0"
-          d="M459.378 273.843c.028-.115.171-.343.543-.686.315-.343.772-.8 1.373-1.373.829.944 1.543 1.887 2.03 2.716.457.83.771 1.601.943 2.316-1.314-.286-2.316-.8-3.116-1.23-.83-.457-1.46-1.143-1.773-1.743z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          fill="#e8ccba"
-          d="M444.997 260.804c-.629 1-1.571 1.744-2.715 2.573-1.172.743-3.43.972-4.346 1.915-1.029.886-.8 2.23-1.63 3.373-.943 1.03-2.8 2.144-3.802 3.116-1.085.886-1.886 1.572-2.287 2.716-.513 1.115-.657 2.944-.428 3.945.172.829.714 1.401 1.63 1.486.857.116 2.258 0 3.544-.8 1.173-.943 2.545-2.944 3.659-4.345.972-1.487 2.344-3.888 2.573-3.802.142.114-.6 2.973-1.343 4.487-.886 1.43-1.888 2.945-3.002 4.202-1.058 1.202-2.058 2.116-3.116 2.86-1.143.657-2.63.857-3.259 1.486s-.657 1.658-.4 2.43c.143.744.915 1.83 1.773 2.173.857.285 2.058.057 3.373-.258 1.2-.4 2.515-.829 3.945-2.172 1.429-1.458 3.544-5.46 4.487-6.374.773-.944.773-.057.544.657-.315.572-1.258 1.915-2.058 3.145-.83 1.258-1.773 2.658-2.83 4.316-.058.972.2 1.658.657 2.173.457.514 1.258.858 2.058.685.715-.371 1.487-1.142 2.431-2.315.885-1.2 2.143-3.173 3-4.603.745-1.6 1.287-3.63 1.745-4.202.486-.572 1.143 0 1.086.943-.229.83-1.315 2.745-2.03 4.203-.744 1.343-1.43 2.744-2.172 4.087a4.125 4.125 0 002.572 1.486c.972.057 2.2-.056 3.26-.685.914-.686 2.029-1.772 2.572-3.259.372-1.63.4-4.43.543-6.089.172-1.715-.028-2.601.83-4.087.828-1.572 2.315-3.373 4.345-5.288-1.915-2.602-3.773-4.717-5.575-6.233-1.858-1.63-3.658-2.372-5.289-3-1.6-.688-3.087-1.002-4.345-.945zM331.966 272.866c.457.858.857 1.915 1.2 3.402.257 1.458.6 3.4.686 5.174-.057 1.744-.114 3.573-.286 5.145-.143 1.401-.571 2.459-.4 3.517.086 1.029.543 1.858 1.372 2.6.8.686 1.973 1.144 3.517 1.487.142-3.088.514-5.575 1.229-7.89.686-2.287 1.972-5.06 2.973-5.689.943-.629 1.886.286 2.715 1.63.658 1.2 1.43 3.917 1.773 5.831.343 1.887.514 3.746.4 5.575 1.2.343 2.287.514 3.259.4.915-.171 2.172-.343 2.573-.943.286-.714.2-1.63-.258-2.858-.457-1.258-1.486-3.345-2.315-4.489-.886-1.257-1.915-2-2.973-2.572.143-1.229-.086-2.572-.686-3.917-.658-1.429-2.459-3.43-2.973-4.345-.658-1-.715-1.715-.286-1.372.457.257 2.315 1.63 3.259 3.116.829 1.486 1.543 3.23 2.03 5.574 1.343 1.602 2.486 2.83 3.544 3.802.972.915 1.658 1.43 2.716 1.773 1.029.285 3.202.656 3.659 0 .343-.773-.315-2.66-1.086-4.346-.887-1.743-2.545-4.03-3.545-5.974-.943-2.058-1.03-4.26-2.03-5.717-1.172-1.43-3.287-1.601-4.345-2.83-.972-1.286-1.658-2.83-1.744-4.746-2.4.2-4.716.972-7.06 2.43-2.316 1.458-4.66 3.46-6.918 6.232z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          fill="#e0e0e0"
-          d="M327.477 279.274c.486-1.543 1.029-2.687 1.63-3.545.57-.857 1.2-1.4 1.886-1.63.343.344.657.801.829 1.373.114.515.2 1.144.143 1.887-.743.572-1.43 1.03-2.173 1.373s-1.514.485-2.315.542z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          fill="#77605b"
-          d="M339.454 301c-.057-3.173-.057-5.86 0-8.176-.028-2.401-.057-4.088.372-5.832.429-1.83 1.515-4.317 2.373-4.83.772-.515 1.63-.345 2.344 1.742.6 2.087 1.2 7.433 1.486 10.406.229 2.887.229 5.002 0 6.689h-2.23c.23-1.172.23-2.087.115-2.859-.257-.886-.629-1.458-1.115-1.601-.514-.085-1.458.429-1.715 1.229-.372.715-.458 1.715-.257 3.117-.457.03-.915.058-1.373.115zM330.65 304.343a40.47 40.47 0 017.204-1.344c2.687-.4 5.717-.571 8.661-.513 2.888.084 5.518.457 8.319 1.114-.515 6.29-1.058 13.75-1.63 23.041-.714 9.262-1.515 24.241-1.972 31.931-.429 7.461-.486 8.29-.629 13.179-.257 4.916-.543 10.176-.857 16.123-3.774.285-7.204.343-10.406.485-3.259.057-6.089.057-8.805 0-.771-13.72-1.229-25.356-1.486-35.561a595.855 595.855 0 010-25.041c-.057-5.232-.429-2.316-.114-6.146.257-3.917.829-9.636 1.715-17.268z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          d="M341.77 306.201c-.972 6.975-1.715 14.894-2.344 24.1-.772 9.147-1.344 21.21-1.487 30.557-.114 9.348.143 17.495.887 24.985.257-19.21.686-35.076 1.172-48.483.458-13.35 1.086-23.584 1.772-31.159z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          fill="#294b70"
-          d="M390.683 302.397c-1.286 1.916-2.716 3.66-4.288 5.117-1.63 1.515-2.944 2.859-5.374 3.688-2.602.658-5.775.344-9.405.486-3.602.057-9.49-.2-12.035 0-2.63.086-3.745.4-3.459.829-.743 6.233-1.258 12.12-1.658 18.124-.457 5.946-.515 11.663-.83 17.38-.37 5.632-.77 10.92-1.2 16.209l-.114.143c1.286 7.032 2.373 12.692 3.316 17.266s1.515 7.318 2.287 9.891c.715 2.573 1.4 4.117 2.144 5.375.715 1.23 1.058 1.572 2.258 2.03 1.172.457 3.716.429 4.774.714 1.03.257 1.601.657 1.43.83-.2.17-1.516.228-2.516.228-1 .028-2.144 0-3.46-.115.687 3.144 1.316 6.49 1.916 10.262.6 3.745 1.23 7.547 1.658 12.035s.686 10.234.972 14.665c.257 4.431.572 7.92.715 11.664.086 3.773-.143 7.575 0 10.748.114 3.145.543 6.203.714 7.976.115 1.715.172 2.515.115 2.372 1.8.115 3.659.143 5.603.115 1.915 0 3.659-.086 5.946-.229s5.889-.343 7.632-.6c1.716-.229 1.916-.258 2.602-.83.686-.6 1.172-1.429 1.43-2.63-.03-2.316-.058-6.632-.115-13.207-.029-6.632.028-15.665-.115-25.985-.143-10.348-.629-26.3-.714-35.476-.086-9.119.085-13.121.228-18.924.143-5.833.6-11.693.6-15.695-.028-4.002-.257-6.66-.714-8.119-.543-.115-1.258-.2-2.144-.228-.886-.057-2.658.143-3.087 0s.057-.715.6-.83c.543-.142 1.687.058 2.601.115s1.83.2 2.86.229c1.057.057 2.858-.085 3.344 0 .486.115 0 .457-.486.6-.458.144-1.258.229-2.373.229.372 3.373.543 8.176.486 14.665-.086 6.518-.772 16.408-.829 24.07-.086 7.604.314 13.55.457 21.326.143 7.803.286 18.38.372 25.013.057 6.574.086 8.948.114 14.293.029 5.374.029 11.149 0 17.61a31.829 31.829 0 014.288.114c1.573.144 3.116.715 5.117.715 2.001-.03 4.26-.572 6.804-.83 2.543-.257 5.26-.514 8.319-.714-.858-10.234-1.287-18.982-1.287-26.7-.028-7.69.658-13.979 1.173-19.181.514-5.203 1.258-8.376 1.914-11.778.63-3.401 1.258-6.174 1.916-8.46-1.03-.116-1.972-.116-2.859 0-.915.084-2.115.713-2.515.6-.4-.172-.486-1.144.114-1.43s2.344-.115 3.459-.257 2.2.6 3.23-.572c1-1.201 1.915-3.088 2.744-6.461.772-3.373.972-5.545 2.001-13.579 1.058-8.09 3.06-26.1 4.174-34.275 1.115-8.119 1.515-8.919 2.516-14.179a1130.36 1130.36 0 013.316-17.152c-3.63.486-7.318.858-11.177 1.086-3.917.2-7.89.372-11.921.23-4.059-.144-8.718-.315-12.15-1.087-3.458-.743-6.174-1.887-8.347-3.43-.943-1.373-1.63-2.43-2.086-3.26 0-.028-.029-.057-.029-.057-.029-.057-.057-.085-.086-.142s-.028-.086-.057-.116c-.457-.885-.629-1.372-.486-1.571l-.343-.943z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          fill="#77605b"
-          d="M388.77 471.828c-.086.314-.715.943-1.773 1.772-1.086.772-2.601 1.858-4.66 3.145.572 1.343.486 2.63-.143 3.973-.743 1.172-1.658 1.858-3.945 3.83-2.4 2.002-7.06 6.662-10.005 8.348-2.944 1.572-6.803 2.63-7.404 1.515-.686-1.201 1.487-5.317 3.43-8.061 1.802-2.859 5.69-6.575 7.805-8.49 2.087-1.972 3.687-2.945 4.802-3.03-.428-.371-.6-.686-.829-.943-.143-.285-.343-.514-.286-.685-2.23.943-4.63 2.715-7.518 5.202-3.001 2.459-6.86 6.29-9.576 9.72-2.687 3.458-5.26 7.49-6.175 10.663-1.086 3.087-.943 6.632.429 7.66 1.343.886 4.574-.4 7.804-2.316 3.144-2.058 7.747-6.689 11.063-9.576 3.202-2.944 5.689-5.088 8.233-7.66 2.43-2.66 5.66-6.575 6.832-7.662 1.2-1.057.858-.057.286 1.23-.658 1.257-2.116 3.115-4.26 6.004l-.4 3.83c1.658-1.659 3.116-3.06 4.374-4.632 1.286-1.571 2.516-2.715 3.145-4.516.457-1.858.2-4.774 0-6.318-.258-1.574-.63-2.574-1.23-3.003z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          fill="#77605b"
-          d="M378.36 475.115c.572-.172 1.43.057 1.916.686.486.6 1.63 1.915 1.115 3.288-.715 1.286-2.487 2.772-4.66 4.774-2.287 2-5.831 5.403-8.204 7.118-2.459 1.572-4.631 2.458-5.632 2.344-1-.2-1.143-1.4 0-3.288 1.058-1.972 4.317-5.803 6.575-8.09 2.259-2.315 5.117-4.516 6.718-5.602 1.43-1.115 1.487-1.23 2.173-1.23zM398.775 468c-.658.43-1.201 1.258-1.516 2.602-.371 1.343-.513 3.23-.543 5.203-.057 1.801-.143 4.46.4 6.004.458 1.515 1.03 1.915 2.488 3.173 1.372 1.172 3.287 2.458 5.86 3.973l.829-3.573c-1.601-.944-2.973-1.715-4.231-2.744-1.286-1.03-3.373-2.888-3.287-3.145.085-.344 1.372 0 3.544 1.372 2.087 1.23 5.832 4.146 8.92 6.433 3 2.315 5.83 5.174 9.033 7.26 3.144 1.916 6.86 3.69 9.718 4.775 2.745 1.143 5.633 2.115 7.119 1.802 1.572-.458 2.2-2.03 1.772-4.26-.63-2.344-2.916-6.174-5.203-9.176-2.459-3.06-5.632-6.233-8.633-8.748-3.03-2.63-6.404-5.032-8.89-6.432-2.46-1.43-4.374-2.03-5.746-1.916.343.486.514.944.4 1.373-.115.428-.428.8-.943 1.23 2.63.829 5.146 1.83 7.661 3.43 2.516 1.543 5.632 3.516 7.376 5.603 1.63 2.03 2.23 4.517 2.343 6.145-.028 1.487-.028 3.173-2.2 2.745-2.259-.6-8.005-4.117-10.95-6.032-2.972-1.973-5.202-3.86-6.688-5.317s-2.344-2.6-2.344-3.43c-1.544.085-2.745-.086-3.831-.829-1.172-.773-2.144-2.172-2.6-3.43-.43-1.26-.344-2.661.142-4.09z"
-          strokeWidth={0.02858652}
-        />
-        <path
-          fill="#77605b"
-          d="M406.834 476.49c.372 1.2 1.83 2.858 3.717 4.516 1.886 1.571 4.86 3.43 7.375 4.916 2.4 1.4 5.574 3.402 6.975 3.573 1.258.057 1.286-1.6.972-2.887-.429-1.458-.915-3.259-2.887-5.06-2.03-1.83-6.233-4.316-8.749-5.603-2.487-1.401-4.86-2.287-6.031-2.201-1.258.115-1.829 1.43-1.372 2.745zM381.39 491.124l-3.829.286 3.944-4.114h0zM406.715 488.956l3.145-.69-2.343-2.2h0z"
-          strokeWidth={0.02858652}
-        />
-      </g>
-      <path
-        d="M554.286 444.133c-1.165-.263.72-2.04.575-3.167 9.54.058 18.53-7.118 29.09-6.626-.216 1.296-1.68.804-2.592 1.152-2.148.804-3.852 1.644-6.912 2.304-6.949 1.501-13.454 4.25-20.16 6.337z"
-        fill="#d2d4d3"
-      />
-      <path
-        d="M515.115 399.777c-3.708.3-6.24-1.272-9.793-1.728 9.41-2.976 18.723-1.74 27.363-3.456 6.553-3.265 13.032-5.605 21.026-6.625 5.22-.672 9.733-.936 14.977 2.016 7.284 4.092 10.021 11.353 13.826 19.298 1.476 3.096 3.348 6.12 4.608 9.217 2.796 6.876 4.164 13.945 6.337 20.45 1.668 4.978 4.02 9.264 6.048 14.688 1.764 4.717 3.216 9.792 4.896 14.976.769 2.377 1.86 4.513 2.88 6.913 4.152 9.818 6.948 20.078 10.08 29.09 1.44 4.154 3.96 7.658 4.32 11.81-2.604.912-3.66-2.317-4.608-4.32-7.512-15.878-13.369-36.856-19.584-51.844-5.977-10.538-10.274-24.723-14.977-36.579-1.44-3.636-3.42-6.866-4.897-10.37-.612-1.451-.54-3.108-1.152-4.608-2.364-5.845-6.732-10.465-10.657-14.978-3.144-.48-7.464-4.044-10.08-1.44 1.344 1.584 3.42 2.388 4.321 4.032 4.872 8.881-.612 25.61-3.744 35.14-1.908 5.821-4.453 10.944-6.049 16.706-2.388 8.605-5.173 16.573-7.488 25.057-4.476 16.381-9.828 34.61-14.69 49.83-.264 6.325-3.456 13.164-4.896 19.584-.6 2.688-.972 4.717-2.016 7.776-.972 2.833-1.404 6.445-4.031 6.337-.492-6.469 2.052-12.77 3.456-19.298 1.212-5.664 2.435-11.437 4.031-17.28 2.676-9.78 4.97-19.875 8.066-29.378 6.504-19.971 10.56-39.665 17.57-59.91 2.7-7.8 5.627-15.733 6.624-24.193.84-7.201-2.172-12.482-8.64-12.962-1.044.012-.744 1.369-1.44 1.729-7.236 1.512-13.68 1.032-20.161 2.016-3.952.6-7.722 2.004-11.526 2.304zM463.554 409.57c.912-3.096-1.188-3.889-2.304-5.474-1.669-2.364-2.22-5.88-2.88-8.353 2.664-.444 1.643 1.033 2.303 2.593 3.313 4.344 6.517 12.013 8.93 17.857 4.044 9.817 8.532 19.729 12.096 30.53.936 2.832 1.608 5.653 2.592 8.353.913 2.497 2.627 4.476 3.745 6.913 1.33 2.892 1.668 6.3 2.879 9.217 1.261 3.048 3.025 5.496 3.744 8.64-.048.336.06.518.288.575-.216.65-1.62.108-2.304.288-2.148-4.416-3.456-8.976-5.184-13.536-1.873-4.932-5.089-9.266-7.201-14.113-2.424-5.557-3.78-11.989-5.473-17.858-2.627-9.084-7.043-17.568-11.231-25.632zM431.583 414.178c.228-.636.6-1.128.576-2.016 8.749-1.333 20.27-6.05 27.938-5.184-4.872 2.412-10.5 2.675-15.84 4.03-3.41.866-7.694 2.389-11.81 3.17h-.864z"
-        fill="#d2d4d3"
-      />
-      <path
-        d="M429.279 377.887c-3.888-1.104-7.729-2.256-11.233-3.744 3.336-1.26 7.513-1.008 11.52-1.728 11.762-2.1 22.25-6.697 32.26-11.233 1.428-.204 1.536-.228 2.592.288 5.112-4.069 11.077-6.409 15.841-10.657 5.629-5.029 8.437-12.145 12.674-18.722 1.813-4.764 3.649-9.732 5.473-15.265.768-2.316 2.352-4.248 3.456-6.625 2.774-5.94 4.2-12.445 6.337-18.433 2.616-7.309 6.996-14.425 10.945-21.026 3.384-5.652 6.396-11.665 12.673-14.689 4.344-2.1 11.605-2.388 17.57-2.88 6-.492 12.384-1.62 17.858-1.152 6.348.54 15.841 2.316 20.45 5.184 2.004 1.452 2.244 4.668 3.456 6.913-.528 9.744-2.904 19.465-2.304 29.666.382 6.48 1.63 12.997 1.727 19.586.217 13.789.169 30.938-2.305 44.643-.648 3.6-2.604 8.197-4.608 10.945-3.552 4.86-7.969 7.429-14.113 10.081-1.692.732-3.24 1.98-4.897 2.88-3.66 1.98-8.328 3.013-12.96 3.745-9.578 1.524-16.742 6.48-25.635 7.488-3.827.432-8.209.096-12.385.576-4.283.492-8.64 1.8-12.384 2.017-3.24.192-6.036-.768-9.218-1.152-7.5-.9-15.925-.589-23.904-.865-5.34-.192-11.113-.624-16.706-1.152-6.168-.576-12.529-.144-18.145-1.152-8.485-1.512-16.766-4.2-23.618-7.776-.636-4.717 2.496-8.21 6.912-9.505 3.769 2.76 9.157 3.9 13.537 6.048 3.157-.432 6.3 1.332 10.37 1.728 1.511.144 3.468-.168 5.184 0 4.752.469 9.624 2.113 14.689 2.88 27.47 4.177 61.061 4.297 87.56-4.32 13.116-4.26 26.28-8.016 31.68-19.873 2.906-6.373 5.522-14.113 5.185-21.026-.288-5.88-1.752-13.081-1.727-20.738.048-14.197-1.321-27.938-1.44-42.915-1.97 11.089-.78 25.178-.576 37.155.073 4.668.24 9.373.576 13.825.312 4.14 1.14 8.16 1.44 12.097.6 7.74-2.305 15.661-5.184 21.602-5.533 11.437-21.218 16.201-35.14 19.873-3.372.888-6.648 2.293-9.792 2.88-3.575.66-7.609.66-11.521 1.153-15.277 1.908-34.31 2.472-53.284-.288-6.901-1.008-13.622-2.437-20.45-3.457-7.117-1.068-17.594.264-19.01-6.912-.518.912.502 2.76.574 4.032z"
-        fill="#d47b3a"
-      />
-      <path
-        d="M396.444 511.53c-.732.217-1.632.288-2.016.866-.24.048-.276-.108-.288-.288-.24.048-.276-.108-.288-.289-.048-4.164 3.132-7.439 4.32-11.231.396-1.25.216-2.532.576-3.745.636-2.1 2.029-3.984 2.593-6.048 1.5-5.449 2.232-10.609 3.744-15.553 1.368-4.465 4.008-8.21 5.76-12.673 1.764-4.5 2.893-9.254 4.897-13.827 1.908-4.332 3.6-9.037 5.184-13.537 1.632-4.631 3.516-9.157 4.897-13.825 2.256-7.585 4.164-15.267 6.336-21.89.684-2.088.6-6.024 4.897-5.184-1.453 3.144-2.713 6.54-4.033 10.08-1.308 3.493-3.276 6.925-4.32 10.37-1.248 4.105-1.932 8.28-3.168 12.384-2.28 7.573-6.421 15.35-9.217 23.042-1.5 4.152-2.16 8.352-3.456 12.673-1.585 5.245-4.045 10.31-5.761 15.554-3.66 11.16-5.988 22.694-10.657 33.121z"
-        fill="#d2d4d3"
-      />
-      <path d="M561.199 394.305c7.104 12.817-1.404 31.275-5.474 44.643 9.793-1.512 16.441-5.952 27.075-7.2-2.328-8.641-7.176-18.52-11.232-26.21-.216-.422.228-.71 0-1.153-2.113-4.067-6.985-9.407-10.37-10.08zm-6.912 49.828c6.709-2.087 13.213-4.836 20.162-6.336 3.06-.66 4.764-1.5 6.913-2.304.912-.348 2.376.144 2.592-1.152-10.561-.492-19.55 6.683-29.09 6.626.142 1.127-1.742 2.903-.577 3.166zm-21.602-49.54c-8.64 1.717-17.953.48-27.363 3.457 3.552.456 6.085 2.028 9.793 1.728 3.804-.3 7.574-1.704 11.522-2.304 6.48-.984 12.925-.504 20.162-2.016.696-.36.396-1.716 1.44-1.728 6.468.48 9.48 5.76 8.64 12.96-.997 8.461-3.925 16.394-6.625 24.195-7.009 20.244-11.065 39.938-17.57 59.909-3.096 9.505-5.388 19.597-8.066 29.378-1.596 5.843-2.82 11.617-4.03 17.28-1.405 6.529-3.949 12.83-3.457 19.298 2.627.108 3.06-3.504 4.031-6.336 1.044-3.06 1.416-5.089 2.016-7.777 1.44-6.42 4.633-13.26 4.897-19.585 4.861-15.218 10.214-33.446 14.689-49.829 2.316-8.483 5.1-16.452 7.489-25.057 1.596-5.761 4.14-10.885 6.048-16.705 3.132-9.53 8.617-26.26 3.744-35.14-.9-1.644-2.976-2.449-4.321-4.033 2.617-2.604 6.936.96 10.08 1.44 3.925 4.513 8.293 9.133 10.658 14.978.612 1.5.54 3.157 1.152 4.608 1.476 3.504 3.456 6.733 4.896 10.37 4.704 11.857 9 26.041 14.977 36.58 6.217 14.987 12.073 35.965 19.585 51.842.948 2.004 2.004 5.232 4.608 4.32-.36-4.15-2.879-7.655-4.32-11.808-3.133-9.013-5.929-19.274-10.081-29.09-1.02-2.4-2.112-4.536-2.88-6.913-1.68-5.185-3.131-10.261-4.896-14.976-2.028-5.426-4.38-9.71-6.048-14.69-2.173-6.504-3.54-13.573-6.337-20.45-1.26-3.096-3.13-6.12-4.608-9.216-3.804-7.945-6.54-15.205-13.825-19.298-5.245-2.952-9.758-2.688-14.977-2.016-7.99 1.02-14.47 3.36-21.023 6.625zm-63.078 21.603c-2.412-5.845-5.616-13.514-8.929-17.858-.66-1.56.36-3.036-2.304-2.592.66 2.472 1.212 5.988 2.88 8.353 1.116 1.584 3.217 2.376 2.305 5.473 4.188 8.065 8.604 16.548 11.232 25.633 1.693 5.87 3.049 12.301 5.473 17.859 2.112 4.847 5.328 9.18 7.2 14.113 1.729 4.56 3.037 9.12 5.185 13.536.684-.18 2.088.361 2.304-.288-.228-.059-.336-.239-.288-.575-.72-3.144-2.483-5.592-3.744-8.64-1.212-2.917-1.548-6.325-2.88-9.217-1.117-2.437-2.831-4.417-3.744-6.913-.984-2.7-1.656-5.522-2.592-8.354-3.565-10.802-8.054-20.715-12.098-30.53zm-35.139-6.049c3.336-2.34 6.829-1.848 10.945-2.592 2.508-.456 4.92-2.148 7.777-2.591 1.956-.3 3.984.587 5.76-.865-1.416-2.713-2.112-6.145-3.744-8.641-6.229-.54-10.837-.708-16.13-.864-1.284 2.148-1.74 4.788-2.592 7.488-.912 2.88-3.588 5.569-2.592 8.065h.576zm25.634-3.168c-7.668-.866-19.19 3.85-27.938 5.184.024.887-.348 1.38-.576 2.016h.864c4.116-.78 8.4-2.304 11.809-3.17 5.34-1.354 10.969-1.618 15.841-4.03zm-29.954-26.786c-4.38-2.149-9.77-3.289-13.537-6.049-4.417 1.296-7.549 4.788-6.913 9.505 6.853 3.576 15.133 6.264 23.618 7.776 5.616 1.009 11.977.577 18.146 1.153 5.592.528 11.364.96 16.705 1.152 7.98.276 16.404-.036 23.905.864 3.18.384 5.976 1.344 9.218 1.152 3.744-.216 8.1-1.524 12.384-2.016 4.176-.48 8.557-.144 12.385-.576 8.892-1.008 16.057-5.965 25.634-7.489 4.63-.732 9.3-1.764 12.96-3.744 1.657-.9 3.205-2.149 4.897-2.88 6.145-2.653 10.561-5.221 14.113-10.081 2.005-2.749 3.96-7.345 4.609-10.945 2.473-13.705 2.521-30.855 2.304-44.644-.096-6.589-1.344-13.105-1.727-19.586-.6-10.2 1.775-19.921 2.304-29.666-1.212-2.244-1.452-5.46-3.456-6.913-4.609-2.868-14.101-4.644-20.45-5.184-5.473-.468-11.857.66-17.857 1.152-5.966.492-13.227.78-17.571 2.88-6.275 3.024-9.288 9.037-12.673 14.69-3.948 6.6-8.329 13.717-10.945 21.025-2.136 5.989-3.564 12.493-6.336 18.434-1.105 2.376-2.689 4.308-3.457 6.624-1.824 5.533-3.66 10.501-5.472 15.266-4.237 6.576-7.045 13.693-12.674 18.721-4.765 4.248-10.73 6.589-15.842 10.657-1.056-.516-1.164-.492-2.592-.288-10.009 4.536-20.498 9.133-32.259 11.233-4.008.72-8.184.468-11.52 1.728 3.504 1.488 7.344 2.64 11.232 3.744-.072-1.272-1.092-3.12-.576-4.032 1.416 7.177 11.893 5.844 19.01 6.913 6.828 1.02 13.549 2.448 20.45 3.456 18.973 2.76 38.005 2.196 53.284.288 3.912-.492 7.945-.492 11.52-1.152 3.145-.588 6.423-1.992 9.794-2.88 13.922-3.673 29.606-8.437 35.139-19.874 2.879-5.94 5.784-13.861 5.184-21.602-.3-3.936-1.128-7.957-1.44-12.097-.336-4.452-.504-9.157-.576-13.825-.204-11.977-1.392-26.066.576-37.155.12 14.977 1.488 28.718 1.44 42.915-.025 7.657 1.439 14.858 1.727 20.738.336 6.913-2.279 14.653-5.184 21.026-5.4 11.857-18.566 15.613-31.682 19.874-26.498 8.616-60.089 8.496-87.56 4.32-5.064-.768-9.936-2.412-14.688-2.88-1.716-.168-3.673.144-5.185 0-4.066-.396-7.21-2.16-10.366-1.728zm-35.715 132.203c.384-.577 1.284-.649 2.016-.866 4.668-10.427 6.997-21.96 10.657-33.122 1.716-5.244 4.176-10.31 5.76-15.554 1.296-4.321 1.956-8.522 3.457-12.673 2.796-7.692 6.936-15.47 9.216-23.042 1.236-4.104 1.92-8.28 3.169-12.385 1.044-3.444 3.012-6.875 4.32-10.369 1.32-3.54 2.58-6.936 4.032-10.08-4.296-.84-4.212 3.096-4.896 5.184-2.172 6.624-4.08 14.305-6.337 21.89-1.38 4.668-3.264 9.193-4.896 13.825-1.584 4.5-3.276 9.205-5.184 13.537-2.005 4.572-3.133 9.326-4.897 13.826-1.752 4.464-4.392 8.21-5.76 12.673-1.512 4.945-2.245 10.105-3.745 15.554-.564 2.064-1.956 3.948-2.592 6.048-.36 1.212-.18 2.495-.576 3.745-1.188 3.792-4.368 7.068-4.32 11.231.012.18.048.336.288.288.012.182.048.338.288.29zm11.233-23.618c-2.448 8.255-3.853 16.884-7.777 24.77-2.544.048-3.768 1.368-6.048.286-1.26-6.263 4.236-10.212 4.608-16.705 4.356-7.931 3.816-17.473 8.065-26.498 3.336-7.08 6.492-15.144 9.792-23.329 2.053-5.088 3.133-10.717 5.185-15.84.564-1.417 1.74-2.617 2.304-4.034 4.32-10.896 5.749-22.548 10.08-33.41-4.595-1.955-11.172-2.447-15.84-4.031-5.317-1.813-10.79-4.129-8.641-11.233 2.604-8.63 13.729-7.777 22.466-9.505 13.8-2.724 26.09-9.073 38.019-12.961 13.201-7.453 21.866-21.422 27.074-37.731 1.272-3.973 3.096-7.309 4.609-10.945 4.764-11.461 7.476-23.402 13.537-33.411.984-1.62 2.208-3.036 3.168-4.608 3.444-5.665 6.47-10.933 11.809-14.978 4.38-2.244 9.074-2.772 14.114-4.608 7.932 1.284 17.785-1.944 27.363-.864 1.14.132 2.292.876 3.744 1.152 1.535.288 3.24-.168 4.896 0 9.613.96 15.506 6.036 15.266 15.553-.026.936.324 2.004.288 2.88-.24 5.269-1.849 11.173-2.305 16.13-.78 8.46 1.8 19.13 1.727 29.954-.048 7.753-.012 14.461 0 22.466.012 10.453-1.464 23.282-5.184 30.243-5.04 9.408-15.421 13.873-24.77 18.145 6.864 2.364 13.369 6.385 16.993 12.673 1.524 2.64 2.388 6 3.745 9.217 1.212 2.88 3.06 5.76 4.32 8.64 2.736 6.241 3.876 13.202 6.913 18.433.217.888-.995.35-.575 1.441 6.024 13.584 10.116 27.661 15.265 40.9 3.78 9.697 6.54 20.244 11.233 30.818 1.379 3.12 3.384 6.517 2.879 10.081-1.679.683-2.22.971-4.32 1.44-1.14-.3-1.165-1.717-2.593-1.728-8.94-19.597-15.157-40.9-23.906-60.774-2.952-6.697-5.675-13.393-8.064-20.162-2.88.252-4.969 2.568-7.489 3.168-4.212.997-8.1 1.093-11.809 2.305-4.008 1.308-7.526 3.792-12.385 3.745-5.16 17.76-11.436 35.127-16.128 53.283-.973 3.745-1.97 7.513-3.17 11.233-1.104 3.397-2.724 6.96-3.456 10.37-.564 2.578-.683 4.978-1.44 7.487-1.427 4.788-2.82 9.854-4.031 14.977-1.248 5.28-2.726 10.705-5.474 15.553-2.687.71-6.168.432-7.2-2.016.42-1.379 1.417-.396 2.304 0 .348-4.248.9-9.3 2.017-14.114 4.165-17.989 9.936-37.334 14.976-55.013.804-2.808 1.176-5.82 2.018-8.64.768-2.569 2.243-4.932 2.879-7.488 3.325-13.357 7.164-26.246 10.945-38.883 1.284-4.297 3.408-8.293 4.32-12.385.6-2.713.636-5.34 1.153-7.776.96-4.512 3.395-11.965-2.017-16.13-6.865 2.017-14.053 2.017-21.602 2.593-5.796.444-10.957 3.096-16.417 2.88-3.66-.144-6.997-2.124-10.657-2.304-.926-.228-1.332-.972-2.018-1.44-1.03-.252-2.23.816-2.59.288-11.415-1.776-24.159-2.292-34.563-2.304 3.288 6.792 6.504 16.153 10.656 23.042.192 3.372 1.669 6.672 2.88 9.793 2.426 6.227 5.341 12.443 7.49 19.01.995 3.046 1.883 6.396 3.168 9.216.973 2.124 2.28 3.996 3.169 6.335 1.596 4.214 3.36 8.162 5.183 12.963.732 1.943 2.27 4.956 0 6.912-1.235-.096-3.887-.552-6.048-1.439.025-2.713-.564-4.813-2.016-6.05-1.717-9.131-7.837-16.98-10.945-26.786-4.32-13.632-8.953-28.274-15.842-39.458-9.624 1.86-20.473 4.68-30.242 6.913-2.172 8.496-5.293 16.477-8.353 24.482-1.5 3.924-3.66 7.62-4.896 11.52-1.308 4.128-2.1 8.281-3.457 12.386-2.616 7.92-5.928 15.72-8.352 23.905z" />
-    </svg>
-  );
-}
-
-export default SvgOfcWork;
