--- src/iconComponents/Slamdunk.tsx
@@ -1,341 +1,0 @@
-import * as React from 'react';
-
-function SvgSlamdunk(props: React.SVGProps<SVGSVGElement>) {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      viewBox="0 0 254.39 504.76"
-      {...props}
-    >
-      <path
-        d="M86.522 24.3c0 13.423-10.653 24.303-23.793 24.303s-23.793-10.88-23.793-24.302c0-13.423 10.652-24.303 23.793-24.303S86.522 10.88 86.522 24.301z"
-        fill="#b64e17"
-      />
-      <path
-        d="M54.31 1.752c-14.419 8.776-16.695 23.421-9.8 38.179-9.825-12.502-7.682-30.833 9.8-38.179zM78.47 9.544c2.52 4.208 3.942 25.891 5.702 14.906C85.933 13.465 75.95 5.337 78.47 9.544z"
-        fill="#ddbdb0"
-      />
-      <path
-        d="M52.2 45.23C33.907 16.706 60.751.545 60.571.429 60.792.313 37.56 16.706 52.2 45.23zM62.85 47.29C59.13 18.146 61.325.745 61.145.629c.22-.116 2.256 18.603 1.705 46.661zM76.83 42.95C84.891 12.101 61.973.435 61.793.319 62.013.203 88.078 9.93 76.83 42.95z"
-        fill="#371a12"
-      />
-      <path
-        d="M85.4 23.99c.842-.168 1.178-.842 2.02 1.347.842 2.188 1.683 8.08.842 12.457-.842 4.376-.337 8.753-1.684 10.269-1.346 1.515-7.238 6.733-8.417 10.437-1.178 3.703-1.515 48.986-2.188 53.026-.673 4.04-3.367 11.952-3.703 15.319-.337 3.366-.337 5.891-.337 5.891s-8.754-1.178-12.625-.168c-3.693.964-6.209.88-6.209.88s.822-22.932 1.496-28.656c.673-5.723 9.426-35.519 9.931-37.876.505-2.356 1.59-11.758 1.852-12.625 1.713-5.657 4.637-5.96 4.882-7.406.402-2.377-1.347-4.377-3.198-6.734s-2.862-2.357-1.852-4.545 2.862-2.357 3.872-1.347c1.01 1.01 3.532 5.626 5.89 7.141s2.695 1.108 4.715-1.754 5.05-11.279 4.545-12.625c-.505-1.347.505-2.862.168-3.03z"
-        fill="#a67054"
-      />
-      <path
-        d="M57.68 131.83c-3.02-2.677-3.246-15.869-2.481-20.105.618-3.42 1.847-6.642 2.173-12.784.36-6.783 7.462-25.454 8.443-29.034.971-3.545.926-17.427 2.337-19.32 1.256-1.683-2.749 3.464-2.543 5.112.206 1.647-.824 8.854-1.03 10.708-.206 1.853-6.589 25.328-7 26.769-.413 1.441-3.501 10.708-3.295 12.56.205 1.854-1.236 26.976-1.236 27.8 0 .823 4.838-2.117 4.632-1.705z"
-        fill="#ddbdb0"
-      />
-      <path
-        d="M55.73 131.88c.618-3.707 4.626-9.472 7.413-10.502 2.788-1.03 2.214 7.955 4.48 9.19s3.096 3.035.007 3.035-10.459-1.723-11.9-1.723z"
-        fill="#ddbdb0"
-      />
-      <path
-        fill="#371a12"
-        d="M104.32 380.31c1.349 1.349 4.497 6.295 6.745 8.543s10.117 8.094 14.614 8.994c4.496.899 41.144 29.453 43.392 32.376 2.248 2.922 17.762 23.607 18.66 23.832.9.225 12.142-14.84 12.142-14.84s-20.46-23.831-22.258-26.08c-1.799-2.248-26.53-29.677-30.577-31.25-4.047-1.574-20.01-17.762-22.033-17.986-2.024-.225-20.235 16.188-20.684 16.413z"
-      />
-      <path
-        d="M121.88 375.42c1.799 6.745 11.466 15.963 13.265 15.963 1.799 0 6.745-2.473 5.396-4.721s-13.94-14.614-15.963-13.715c-2.024.899-2.024 2.698-2.698 2.473zM174.72 419.04c.666-.256 7.001 8.689 8.125 7.34 1.124-1.35 3.22-2.868 2.321-4.217-.9-1.349-7.305-9.401-9.104-9.85-1.799-.45-2.008 6.983-1.342 6.727z"
-        fill="#a67054"
-      />
-      <path
-        d="M202.05 439.08c2.023 2.248 2.995 7.547 4.57 8.672s15.287 17.986 15.063 21.584c-.225 3.597-12.815 34.624-14.839 35.298-2.023.674-5.396-1.574-6.52-3.148s.9-12.141.225-13.49-5.396-8.543-6.07-11.242c-.675-2.698-3.598-12.366-5.846-16.862-2.248-4.497-5.846-5.396-3.822-5.396s4.721-.9 6.52-2.473c1.799-1.574 10.944-13.393 10.719-12.943z"
-        fill="#d9ae78"
-      />
-      <path
-        fill="#6f3719"
-        d="M207.31 439.71c.948-1.896 13.27-11.849 13.27-14.455 0-2.607-21.327-26.541-24.171-29.858-2.844-3.318-19.195-21.564-24.645-24.408-5.45-2.844-14.218-7.583-17.536-9.005-3.318-1.422-11.138-11.138-12.086-11.612-.948-.474-19.906 21.327-20.616 22.75-.71 1.42 8.768 6.871 10.427 9.478 1.659 2.607 20.143 6.635 24.645 8.768 4.502 2.132 26.778 22.038 29.858 25.119 3.08 3.08 18.958 22.986 20.853 23.223z"
-      />
-      <path
-        d="M128.78 361.02c0 3.555 10.123 20.572 15.81 20.098 5.687-.474 10.118.54 12.725 3.858 2.607 3.318 30.806 20.616 35.072 26.778 4.266 6.161 19.906 23.697 19.906 23.697l-4.74 4.74s-18.246-22.512-21.563-25.12c-3.318-2.606-19.432-17.535-22.512-19.194-3.081-1.66-4.266-3.792-7.11-4.266-2.843-.474-19.194-5.45-22.037-6.635-2.844-1.185-11.375-9.479-12.322-10.901-.948-1.422 6.773-13.056 6.773-13.056z"
-        fill="#371a12"
-      />
-      <path
-        fill="#e6e6e6"
-        d="M100.75 267.63c2.978.993 10.918 3.226 13.4 5.955 2.481 2.73 8.933 9.43 9.43 11.663.496 2.234 1.488 4.715.496 3.97-.993-.744-2.978 0-1.241 3.475s3.722 3.225 4.466 7.196c.745 3.97 0 4.962 3.474 6.203s9.926 6.7 10.67 9.678c.745 2.977 1.737 10.422 2.978 14.144 1.24 3.722 5.46 4.714 6.452 4.466.992-.248-2.234 6.204-3.474 8.685s-24.07 28.288-25.807 30.273c-1.737 1.985-4.715 1.737-4.715 1.737s-6.7 3.474-9.926 3.474-27.295 6.948-29.777 6.452c-2.481-.497-3.474-6.7-4.714-11.166-1.24-4.467-5.46-24.07-7.444-30.025-1.986-5.956-2.482-13.151-3.226-18.61-.745-5.46-2.73-9.181-3.97-12.655-1.241-3.474-3.226-6.948-2.978-11.414.248-4.467 41.687-32.258 42.928-33.003 1.24-.745 3.226.744 3.226.744l-.248-1.24z"
-      />
-      <path
-        d="M81.32 165.11c2.144.179 2.68-1.072 5.181 1.25s4.824 5.718 5.36 10.542c.536 4.823.715 8.933 1.786 13.042 1.072 4.109-.178 2.322 1.787 5.895s2.323 10.362 2.323 12.863c0 2.502.178 6.61-.536 11.256-.715 4.646.357 11.256-2.501 14.471-2.86 3.216-5.003 5.181-5.182 6.79-.178 1.607 1.966.178 3.574.893s2.144.893 2.68 2.501c.536 1.608 4.109 3.394 3.93 5.002-.179 1.608.536 4.467-.715 5.718s-.178 6.431-1.965 7.682c-1.786 1.25-1.25 1.787 0 2.68s2.859.893 3.216 2.144c.357 1.25-23.404 14.47-26.442 16.794-3.037 2.322-14.47 10.005-16.437 12.863-1.965 2.858-2.144 3.573-2.144 3.573s-.893-3.752-2.501-5.36c-1.608-1.608-2.859-1.608-3.395-2.858-.535-1.251-11.792-27.692-13.22-29.48-1.43-1.786-5.896-7.146-7.504-11.076-1.608-3.93-2.144-16.615-2.323-19.831-.178-3.216-.357-10.362-.714-13.4-.358-3.037-.18-5.538.893-8.396s8.04-15.186 9.29-17.151c1.25-1.966 5.896-8.219 6.61-10.005.715-1.787 3.752-5.539 4.11-6.254.357-.714 29.836-13.22 34.839-12.149z"
-        fill="#d9ae78"
-      />
-      <path
-        d="M47.37 177.44c.536.893.715 5.002-1.072 7.325-1.787 2.323-5.717 7.682-6.968 9.648-1.25 1.965-.714 1.072-.714 1.072l1.072 1.786s4.466-7.146 8.933-8.754 23.94-7.147 25.727-10.541c1.786-3.395 1.429-6.61 2.322-6.968.894-.357 5.36 1.608 5.896 3.216.536 1.608 1.072 8.04-3.573 9.29s-19.831 2.68-21.082 5.36c-1.25 2.68-2.323 5.181-2.323 5.181s-8.754 2.323-11.97 6.79c-3.216 4.466-4.824 9.29-5.717 10.361-.893 1.072-5.002-4.287-5.896 1.787-.893 6.074-5.281-.181-5.18-2.144.096-1.888 17.956-30.835 20.545-33.409z"
-        fill="#f0f4f5"
-      />
-      <path
-        d="M46.35 140.98c1.01-1.6 2.862-4.377 6.733-7.407 3.872-3.03 15.655-2.357 19.19-.673 3.536 1.683 11.785 5.892 12.626 9.932.841 4.04 1.178 11.784 0 14.477-1.179 2.693-7.576 11.447-8.08 14.645-.506 3.198 1.851 3.703-.674 6.228s-17.002 6.397-22.052 9.09c-5.05 2.694-7.744 4.209-10.27 5.892-2.524 1.684-4.376 4.882-5.05 2.694-.673-2.189-.056-.786 2.132-4.49s4.18-4.49 5.695-7.351c1.515-2.862 1.6-3.646 1.432-5.498-.168-1.852-1.515-.842-3.703-1.683-2.188-.842-4.04-1.01-3.703-3.872.336-2.862 0-2.357-1.01-3.535s-1.01-1.01-.169-3.03c.842-2.02 3.03-1.347 2.357-2.526-.673-1.178-3.535-5.723-3.703-7.575-.169-1.851 2.861-1.01 2.861-1.683 0-.674-2.188-2.357-2.02-4.04.169-1.684 4.209-3.199 5.724-5.219 1.515-2.02.673-2.777 1.683-4.376z"
-        fill="#a67054"
-      />
-      <path
-        d="M49.97 149.69c-2.353 3.153-2.936 4.125-3.778 5.69-.774 1.256-.309 2.276 1.071 2.653 1.38.377 1.115-3.594 2.222-1.285s-1.577 5.185-2.813 6.542-4.296 3.994-5.077 4.775c-2.533 1.069.528 1.124.713 2.38.185 1.257-1.825 2.955-1.4 4.276.356 1.108 1.52.842 1.906.313 1.182-1.622-2.226-1.902.187-3.25s2.229-.933 6.969-1.5c7.474-13.334 7.04-4.702 10.406-1.031.067-.313.188-.663.281-1-.146 2.413-.323 4.979-.375 6.125.432.638 1.451.916 2 .375 9.2-6.628-1.817-12.903 1.91-19.64.81-1.659 4.426-3.658 3.59-4.892-1.389-2.05-7.977-3.564-11.685.852-3.708 4.417-5.93 5.748-6.363 1-.432-4.747 10.932-4.965 8.68-5.958-2.253-.992-6.092.422-8.444 3.575zm17 22.688c-4.861 1.482-4.268 2.702 0 0zm5.566-7.116c.848-.173-5.712 7.07-4.476 7.202 2.086-.159 11.233-1.96 10.223-5.211-.673-3.122 3.01-14.386 1.25-12.969-3.164 2.546-3.213 8.564-6.997 10.978z"
-        fill="#6f3719"
-      />
-      <path
-        d="M61.01 184.06c-.524-2.27 2.238-7.129 3.984-8.613s3.756-3.35 6.481-3.94c1.366-.297 4.811-3.392 5.773-2.008.381.55-1.174 3.913-.3 5.048.574.745.839 2.178-.186 3.147-1.295 1.224-14.18 7.064-15.752 6.366zM49.12 147.78c-.256-.016-.53.012-.781.031-1.003.077-3.097 2.028-4.472 2.333-1.375.305-1.57.343-.038.267 1.532-.075 1.511.125 1.694.265.771.771.693 2.802.312 3.104-.579.459 2.166-1.354 2.535-2.219.37-.864 1.415-2.05 1.875-2.844.46-.793-.357-.889-1.125-.937z"
-        fill="#6f3719"
-      />
-      <path
-        d="M63.38 136.23c.463 4.32-4.166.926 0 5.71s2.777 8.11 6.48 8.11 6.4-.176 8.252 1.058 1.007 1.323 1.47 4.255 1.388 5.71 2.16 6.48 4.596-6.136 4.417-10.452c-.18-4.32.523-7.137-2.103-10.995-3.07-4.51-7.56-6.635-11.88-7.407-4.321-.771-6.173-.771-8.179-.463s5.555.31 6.018 2.315-4.166 2.314-6.635 1.389z"
-        fill="#6d5547"
-      />
-      <path
-        d="M70.24 156.48c-1.862.183-2.763-2.562-.24 1.517 1.1 1.777 2.44 1.508 3.733.824 1.294-.685-.076-.761-.532-2.207-.457-1.446-.076-1.522-1.826-.99-1.75.533-.983.78-1.135.856z"
-        fill="#371a12"
-      />
-      <path
-        d="M40.68 153.27c-.372-.548-1.547-1.79-1.064-2 2.17-.945 4.203-.764 4.262.903.075 2.147 2.237 1.539 1.947 1.651-.703.273-1.426.146-2.815.223-1.388.077-1.955-.224-2.33-.777z"
-        fill="#6f3719"
-      />
-      <path
-        fill="#502f20"
-        d="M48.6 145.4c2.54-2.32 1.767-3.534 4.418-2.982 2.65.552 2.982 1.988 3.755 2.54.773.552-4.307.663-5.964.994-1.657.332-2.32-.442-2.209-.552z"
-      />
-      <path
-        d="M50.92 143.14c-.591 1.441-.838 1.959-.591 2.402.246.444 1.38.292 1.861.206.48-.086 3.104-.23 3.14-.477.038-.246-.617-2.313-1.183-2.473-.567-.16-2.02-.607-2.5-.397-.48.21-.752.727-.727.739z"
-        fill="#120500"
-      />
-      <path
-        d="M50.5 142.4c1.003-.926 3.009-.617 5.015.463s1.543 1.345 2.083 1.5c.54.153-.552-.367-1.417-2.014-.973-1.853-2.9-1.552-3.75-1.459-.874.096-2.81.134-3.199 2.255-.389 2.121-1.335 2.36-.824 2.163.511-.197.935-1.442 2.092-2.908z"
-        fill="#150000"
-      />
-      <path
-        d="M51.14 146.3c.953.121 1.06.394 1.713.203.233-.068.504-.702-.961-.548-1.466.155-1.985.037-2.256.453-.27.416.551-.23 1.504-.108z"
-        fill="#ddbdb0"
-      />
-      <path
-        fill="none"
-        d="M49.88 147.88c-.077 4.938-2.623 5.555-2.932 6.79s2.315 3.24 3.395 3.857 3.472-4.475 3.086-6.172c-.386-1.698-1.389-3.472-1.543-4.475-.154-1.003-1.389-.772-1.852-.54s0 .617-.154.54z"
-      />
-      <path
-        d="M38.62 156.88c-.046-.025-.029.1 0 .562.077 1.235 2.629 4.791 2.938 5.563s.928 1.3 1.468.375c.252-.432.827-.679 1.407-.844-.764.557-1.869 1.986-2.44 2.184-1.047.363-2.716 1.478-2.744 2.66-.598 4.247 3.53-.767 4.683-1.906.786-.857 2.076-2.62 1.72-3.188.057-.01.17-.023.187-.031-1.543-1.466-5.139-4.348-6.219-4.656-.675-.193-.924-.677-1-.719z"
-        fill="#4a2318"
-      />
-      <path
-        fill="#1a0301"
-        d="M42.63 163c-1.234-1.389-2.7-2.546-1.774-2.623.925-.077 2.931.694 3.317 1.157.386.463-.694 1.852-1.543 1.466z"
-      />
-      <path
-        d="M43.59 164.37c1.092-1.253 2.542-2.976 2.7-1.848.17 1.216-3.112 3.755-4.035 4.991-.926 1.241-2.61 2.065-2.995 1.332-.23-.437.007-.772.314-.97.307-.198 2.924-2.252 4.016-3.505z"
-        fill="#a67054"
-      />
-      <path
-        d="M39.36 151.11c-.694.077-.382-1.073.32-1.658 2.196-1.702 1.672.674-.32 1.658z"
-        fill="#ddbdb0"
-      />
-      <path
-        d="M40.78 156.21c2.083-1.389 3.395-2.006 2.469-2.083s-1.712.49-2.096.765c-.383.276-1.811.098-1.412.893.417.834 1.039.425 1.039.425z"
-        fill="#6f3719"
-      />
-      <path
-        d="M41.76 153.36c-.463-.391-.237-.89-.113-1.403.124-.513.602-.51 1.166.203.492.623 2.007 1.68.542 1.602-1.466-.077-1.132-.01-1.595-.402z"
-        fill="#1b0403"
-      />
-      <path
-        fill="none"
-        d="M60.61 161.46c.926-2.083 5.632-4.706 6.79-6.404s2.622-2.931.694-2.546c-1.93.386-7.021 1.543-8.487 3.78-1.466 2.238-2.314 1.158-1.774 3.164.54 2.006 1.697 2.7 1.928 3.935.232 1.234.772-1.312.849-1.929z"
-      />
-      <path
-        fill="none"
-        d="M68.32 154.21c-2.546 7.946-.077 10.261-2.006 12.576-1.929 2.314-5.478 6.79-6.712 6.943s-1.466.31-.772-2.468c.695-2.778 0-5.787.618-7.638.617-1.852.694-1.852 3.78-4.475s4.398-4.86 5.092-4.938z"
-      />
-      <path
-        fill="none"
-        d="M64.08 154.82c-1.003 3.78-1.312 4.552.154 3.163s3.55-2.854 3.55-3.703c0-.849.076-1.543-1.621-1.157-1.697.386-1.543.463-2.083 1.697z"
-      />
-      <path
-        d="M79.15 151.41c-.18-.003.005 1.045-.344 3.25-.585 1.572.184 2.98.219 4.5-.112 2.461-1.24 7.52-1.813 9.781-.572 2.262-3.425 1.765-5.53 2.344-4.694 1.252-2.01-1.825.124-4.102s5.222-11.722 2.907-10.742c-2.337.561-3.716-.906-6.094-1.875.047.049-.305 4.21 1.937 6.406 3.708 3.63 1.758 4.54.319 6.593s-6.143 4.494-7.43 6.37c3.197-3.183-1.151-12.63-2.358-13.744-3.165 3.053.695 14.682-2.062 2.438-.44-1.599-1.197-1.414-2.75-3.938s2.352-6.54.25-5.843c-2.102.697-8.353 7.782-8.741 13.953-.247 3.423-5.235 3.896-5.978 5.328-.774 1.49-1.468 2.344-.562 3.25s3.058 1.697 5 1.625c2.008-.075 1.73.617 1.562 3.687-.174 3.185-3.225 5.04-9.312 14.75 7.464-5.969 20.735-10.639 23.156-11.375 2.42-.736.734-6.138 3.25-8.062 2.472-3.835 11.39-3.646 13.28-7.375.307-.862 3.136-5.03 4.063-6.781 1.364-2.677.507-5.817-1.656-8.594-.947-1.215-1.298-1.841-1.438-1.844z"
-        fill="#4a2318"
-      />
-      <path
-        d="M38.49 156.69c-.309-.772-.339-1.329.51-1.174s1.003 1.388 3.24 3.24c2.238 1.852 3.798 3.264 3.49 3.418-.31.154-3.918-3.744-5.693-4.36-1.774-.618-1.316-.815-1.547-1.124z"
-        fill="#a67054"
-      />
-      <path
-        fill="none"
-        d="M48.34 168.02c-.386-2.469-1.697-1.08.309-3.626s3.626-4.32 3.163-2.932-1.775 4.243-2.006 5.092c-.232.849-1.08 1.62-1.466 1.466z"
-      />
-      <path
-        d="M76.32 150.31c2.968 1.37 5.205 1.563 2.234 4.477-3.152 3.092-1.352 2.913-3.63 3.68-2.432.818-3.085.72-5.246-.662-1.434-.917-1.131-3.915-.827-4.676 1.062-2.85 4.654-3.662 7.469-2.819z"
-        fill="#6f3719"
-      />
-      <path
-        fill="none"
-        d="M81.13 161.07c-1.234-4.32.463-6.172 1.389-6.48.926-.31 2.931-1.544 2.314 1.542s-2.006 6.944-2.932 6.018-.308-1.389-.771-1.08z"
-      />
-      <path
-        d="M68.66 153.03c1.27-2.15 2.29-2.854 4.558-3.204 2.268-.35 10.734.577 3.665 6.868.999-1.949 6.262-5.51-2.534-6.121-5.041-.35-4.7 1.798-5.69 2.457z"
-        fill="#a67054"
-      />
-      <path
-        d="M70.39 153.87c-1.41.57-1.926 1.744-1.747 2.352.259.88 2.862-.665 2.634-1.502-.228-.837.523-1.42-.887-.85zM74.53 154.31c-.158 1.782.913 1.698.632.95-.69-1.829 1.403-1.776.837-2.195-.662-.49-1.423.723-1.469 1.245z"
-        fill="#a67054"
-      />
-      <path
-        d="M77.2 150.64c.256.243 1.85.612 1.849 1.383-.003 1.22.88.463-.216-.658-.947-.97-1.89-.968-1.633-.725zM75.93 153.5c-.234-.116-.695.309-.804.802-.108.494-.147 1.075.526.572.674-.502.512-1.258.278-1.374zM69.9 154.68c-.936.707-.764 1.112.306.639.86-.381.847-.792.425-.962s-.462.07-.731.323z"
-        fill="#ddbdb0"
-      />
-      <path
-        d="M62.36 131.96c3.24 4.32 7.04 5.502 9.817 5.656s7.407.154 9.721 2.469 4.287 9.427 4.166 6.942c-.206-4.216-3.241-9.308-6.327-11.16-3.086-1.85-8.166-3.18-11.211-3.702-3.292-.566-8.79-.668-6.166-.205z"
-        fill="#6d5547"
-      />
-      <path
-        d="M77.77 153.79c-.14-.914-.408-1.308-1.754-1.631-2.198-.529-4.655 1.256-4.835 2.36-.255 1.561-3.184 2.082-2.048 1.973 1.023-.097 1.944-1.336 2.803-1.128.827.2.584 2.078 2.182 1.85 1.598-.228 1.657-.614.668-2.44s2.375-2.886 2.68-1.593c.304 1.294.304.38.304.609z"
-        fill="#120100"
-      />
-      <path
-        d="M71.1 137.21c4.424.625 8.26 1.433 10.266 4.21s4.958 3.71 4.752 7.825c-.204 4.054-.413 5.25-.413 5.25s1.134-8.85-3.65-11.165c-4.783-2.314-7.928-4.974-10.86-4.666-2.057.217-6.483-1.128-5.554-2.46.782-1.12 1.035.38 5.459 1.006z"
-        fill="#9b8884"
-      />
-      <path
-        d="M51.92 147.02c.342-.018.55.268.252.414-1.184.586-1.04.998-1.441.883-.421-.12-.675-.94-.397-1.225.28-.286 1.063-.124 1.586-.072zM52.67 150.76c.648-.522 1.386-1.929.115-1.843s-1.85.773-2.048 1.711c-.2.939 1.285.654 1.933.132z"
-        fill="#ddbdb0"
-      />
-      <path
-        fill="#fff"
-        d="M52.78 143.63c1.043.453.549 1.51-.32 1.589-.87.079-1.23.336-1.409-.712-.178-1.047.686-1.33 1.729-.877z"
-      />
-      <path d="M52.95 144.41c.098.861-1.338 1.346-1.923.939-.646-.45-.848-1.084-.289-1.476 1.107-.897 2.114-.324 2.212.537z" />
-      <path
-        d="M71.93 155.29c.74.966 1.114 2.364 2.63 1.657 1.257-.586.459-2.015.459-2.015s1.022 1.322-.084 2.124c-1.045.758-2.668.09-3.229-.05-2.062-.856-1.136-1.485.224-1.716z"
-        fill="#a67054"
-      />
-      <path
-        d="M46.15 143.61c.132-.99.02-1.803 1.148-3.683s4.025-4.387 5.01-5.208c-1.015 1.068-4.767 3.078-4.339 8.178.435 1.208-5.219 4.554-4.68 4.085.726-.631 2.729-2.381 2.86-3.372zM39.36 167.47c.523-1.388.887-1.683 1.66-2.025.774-.34 2.116-1.023 2.39-1.365.272-.34.773-.318.568-.113s-3.503 3.048-3.89 3.457c-.387.41-1.047.728-.728.046zM40.94 175.19c.683-4.466 1.846-3.588 3.562-4.635 1.466-.894 1.566-1.106 1.566-1.106s-5.366-.597-5.307 1.523c.049 1.733-.638 3.046.179 4.218zM69.02 34.66c.11 2.49 7.67 12.588 10.623 10.844 2.953-1.745 3.887-1.52 4.02-3.668.135-2.147-1.073.269-2.818 1.208-1.744.94-5.32-.24-6.844-2.416-1.464-2.087-5.092-8.458-4.981-5.968z"
-        fill="#ddbdb0"
-      />
-      <path
-        d="M73.3 39.62c.412.72 3.634 4.978 6.62 3.846s10.096-9.285 8.14-15.977c-1.957-6.693-3.398-2.883-2.987-1.545.412 1.339-1.132 7.516-1.75 8.855-.618 1.338-3.482 7.068-4.615 7.273-1.132.206-3.76-.805-5.408-2.452z"
-        fill="#4a2318"
-      />
-      <path
-        d="M68.03 33.94c-.084.004-.16.006-.25.031-.72.206-1.956 1.97-1.75 3-.206.31 2.76 4.327 3.688 5.563.926 1.235 1.338 2.876 1.75 3.906.411 1.03-1.548 1.94-2.063 2.969-.488.976-1.077 8.238.031 8.062 1.108-.175-2.436 25.292-7.045 35.5-2.568 5.688-5.768 26.475-5.562 29.77.206 3.295 5.703-11.725 9.036-14.424 3.711-3.006 1.073 12.128 2.309 15.423 1.235 3.295 6.72-10.994 6.513-15.936-.411-4.118-1.568-20.102-.735-26.65.833-6.55 3.066-17.435 3.736-23.573.412-2.265 9.315-8.12 9.457-11.438.078-1.81-5.998 12.098-14.263-2.829-.845-1.526-4.11-7.082-4.42-8.009-.27-.81.153-1.392-.432-1.366z"
-        fill="#6f3719"
-      />
-      <path d="M83.11 165.11c.536 2.144 1.43 1.25 1.787 4.824.357 3.573.992 7.124-.715 9.826-1.685 2.667-7.904 11.223-6.79 13.757.594 1.348-.714 3.037.358 4.645s8.576 8.397 9.112 10.72c1.543 6.516-.862 14.202.714 19.831.18 2.501.715 6.61-.714 7.504s-7.325.715-7.325.715 3.394-9.47 1.965-13.935c-1.43-4.467-3.752-17.151-5.717-19.117-1.965-1.966-10.541-8.04-13.221-8.04-2.68 0-6.253.893-7.147-1.072-.893-1.965-.536-3.395.536-5.181s4.89-4.813 10.541-5.36c5.539-.536 10.005-1.072 12.506-2.68 2.502-1.608 5.182-4.645 3.395-6.253-1.787-1.608-4.645-2.501-5.003-3.752-.357-1.25 2.144-6.432 5.718-6.432z" />
-      <path d="M56.31 195.84c-1.965 0-12.328 3.93-13.935 8.576-1.608 4.645-4.11 4.645-3.93 8.933.178 4.287 3.573 11.792 3.394 12.863-.179 1.072-1.25 8.218-1.25 8.218s1.607-9.112-.537-11.792c-2.143-2.68-3.037-10.541-1.965-13.578 1.072-3.037 3.752-9.29 6.61-11.077 2.86-1.787 10.542-3.93 10.542-3.93l1.072 1.786z" />
-      <path
-        d="M85.08 165.75c.804.268 5.254 7.044 5.79 8.473.536 1.43 1.24 7.322 1.06 10.36-.178 3.037 2.593 11.247 3.673 13.665.92 2.064-3.147-7.113-4.934-5.683s-4.524 7.918-3.452 10.777c1.072 2.858-.536 3.751-1.966 2.5s-7.146-7.146-7.86-8.396c-.715-1.25-.894-3.93.178-6.253s6.968-12.328 7.504-14.83c.536-2.5-.894-9.29-1.608-10.54-.715-1.25.811-.341 1.615-.073z"
-        fill="#f0f4f5"
-      />
-      <path
-        fill="#f9eaaf"
-        d="M96.15 199.06c0 2.68 1.072 9.648-.357 12.328-1.43 2.68-.358-.357-1.608-.536s-.715 1.43-.894 5.538c-.178 4.11 2.68 6.254 1.787 8.576s1.608 5.539-3.216 10.541c-4.824 5.003-9.826 9.47-14.47 9.47-4.646 0-6.79.356-10.006-.716s-29.3-8.933-30.729-9.647c-1.43-.715-8.754-7.861-9.826-10.362-1.072-2.501-1.072-3.752-1.072-3.752s-1.25 6.61 1.25 10.184c2.502 3.573 9.648 6.432 13.221 8.754 3.574 2.323 31.98 7.504 35.375 7.147 3.395-.358 13.221-4.824 14.114-5.896.894-1.072 5.718-5.36 6.432-8.04.715-2.68.715-7.682 1.072-10.184.358-2.5.894-10.898.894-12.506 0-1.608-.358-9.826-1.966-10.898z"
-      />
-      <path d="M86.5 223.53c.926 4.95.4 7.289 1.43 11.256.857 3.308-9.078 5.782-13.043 4.645-3.477-.997-8.397-.536-9.29-.357-.893.178.357-.894.357-.894s3.752-.357 6.61-.714 13.4.536 14.294-1.965c.893-2.502 1.072-10.362-.358-11.97z" />
-      <path
-        fill="#f9eaaf"
-        d="M53.63 296.25c.357-.893 4.466-5.717 6.432-6.968 1.965-1.25 1.965.536 1.965.536s5.538-1.608 8.04-5.002c2.5-3.395 3.216-3.752 3.216-3.752s10.898-6.61 12.685-7.861c1.786-1.25 9.29-7.861 10.362-8.04 1.072-.178 4.109 2.501 4.109 2.501s-26.8 16.258-28.228 17.51c-1.43 1.25-13.221 9.29-14.471 11.612-1.25 2.323-2.501 4.824-2.501 4.824l-1.608-5.36z"
-      />
-      <path
-        fill="#0d0906"
-        d="M71.85 246.4c1.608-.179 3.752.357 3.752.357s-2.323 1.966-4.645 3.216c-2.323 1.25.357.715 2.858 1.787 2.501 1.072 3.93 1.608 6.075-.179s7.86-5.538 10.362-5.002c2.5.536 5.538 1.608 4.287 1.429-1.25-.179-13.042 4.288-16.079 6.074-3.037 1.787-5.895 1.608-5.895 1.608s1.25 2.68 4.823 1.608 5.896-3.93 8.219-5.36c2.322-1.429 5.896-.357 4.645 0s-9.112 1.43-6.79 1.966c2.323.536 4.646 3.037 3.753 2.5-.894-.535-4.11 2.145-7.861 1.966-3.752-.179 1.786 2.323 2.858 1.25s-.178 1.073-.357 1.787c-.179.715 3.573 5.181 3.573 5.896 0 .715-4.824-.893-8.04-1.608-3.216-.714-5.895-.357-7.503-1.25-1.608-.894-1.608-1.787.178-3.93s4.467-2.323.536-2.68c-3.93-.358-6.967 1.964-8.04.357s-2.5-6.79-1.786-8.933c.715-2.144 4.11-3.395 4.11-3.395l6.967.536z"
-      />
-      <path
-        fill="#1a1a1a"
-        d="M52.55 292.14c3.394-2.144 2.501-4.824 3.93-7.861s6.79-5.538 12.329-6.79c5.538-1.25 12.684-1.25 16.614-5.895 3.931-4.645 8.933-5.717 9.47-6.432.535-.714-1.787 3.395-3.395 4.11-1.608.714-13.578 8.754-16.615 10.54-3.037 1.787-6.074 6.968-8.576 8.04-2.5 1.072-3.037.358-6.074 1.965s-6.432 6.79-6.432 6.79-3.216-.715-3.752-2.144c-.536-1.43 2.68-2.323 2.501-2.323z"
-      />
-      <path
-        fill="#f9eaaf"
-        d="M98.11 248.01c-3.573 4.466-4.824 2.68-6.253 4.11s1.608 1.607 4.288 2.322c2.68.714 3.752 1.072 3.752-.536s.178-5.36-.715-5.717c-.893-.358-1.072-.358-1.072-.179zM89.72 241.94c-2.68 2.144-7.146 4.11-8.04 5.36-.893 1.25 6.61-3.573 9.47-2.501s5.538 1.965 5.002.536c-.536-1.43-4.824-3.574-6.432-3.395zM98.29 257.66c-.536 4.824-1.25 4.824-3.573 3.93-2.323-.893-4.467 2.86-5.003.894-.536-1.966.536-5.003 3.573-5.003s5.36 0 5.003.179z"
-      />
-      <path
-        d="M73.95 379.04c0 1.489 2.493 6.17 2.493 6.17.334.253 45.213-7.545 43.143-10.862-.657-1.053-8.415-3.993-8.415-3.993s-31.514 5.956-37.22 8.685z"
-        fill="#d9ae78"
-      />
-      <path
-        d="M39.16 210.67c-1.25 3.573 3.037 12.149 3.037 15.007 0 2.859-2.322 17.687-3.752 21.439-1.429 3.752-.714 13.757-1.072 17.151-.357 3.394-6.253 14.65-7.146 18.045-.893 3.394-.715 14.293-2.501 21.618s-6.79 24.298-7.325 26.084c-.536 1.787-3.216 6.074-5.181 6.968-1.966.893-2.144.178-3.216-.894s-2.144-.178-2.144.894 1.43 2.322 1.965 3.037c.536.715-.893 6.432-3.216 9.111s-7.325 7.861-7.682 9.112c-.357 1.25 2.501 2.859 3.752 1.25 1.25-1.607 5.538-8.396 5.896-9.11.357-.716-10.184 14.113-10.541 15.542-.358 1.43 2.143 2.68 3.93 1.787s8.397-10.898 9.47-12.15c1.071-1.25-.358 6.432-1.609 9.47s-3.773 5.717-2.88 6.431c.894.715 3.773 1.072 5.203-1.25s4.466-10.362 4.466-12.328c0-1.966-.357 8.397-.536 10.54-.178 2.145 1.25 3.753 2.144 2.68.894-1.071 2.859-10.897 2.68-13.934-.179-3.038 3.037-5.717 5.003-8.755s2.144-10.005 2.5-13.22c.358-3.217 16.616-38.234 17.51-40.557.893-2.322 5.18-18.938 5.18-20.546 0-1.608 7.147-11.434 8.22-14.293 1.071-2.858 3.572-13.399 3.572-16.258 1.743-7.196.896-4.93 3.574-5.717 2.677-.786 9.29.179 10.362-.357 1.072-.536 2.68-9.29 3.037-11.97.357-2.68-3.627-12.207-5.003-18.223-1.375-6.016-10.898-9.826-13.22-11.077-2.323-1.25-6.432-1.43-8.576-.357-6.392 2.96-12.18 8.731-15.901 14.829z"
-        fill="#a67054"
-      />
-      <path
-        d="M69.81 223.24c1.18-1.918 3.098 0 3.541 4.131.443 4.131 5.686 9.279 4.118 9.534-1.78.29-9.516.055-10.8.928-1.285.873 2.013-11.88 3.141-14.593z"
-        fill="#371a12"
-      />
-      <path
-        d="M57.42 197.42c1.18 4.426 6.197 8.853 8.853 7.672 2.655-1.18 5.67-3.353 3.675-4.786-2.048-1.47-9.872-4.657-12.528-2.886z"
-        fill="#debfb3"
-      />
-      <path
-        d="M48.86 202.73c.148 2.95-2.656 3.246-4.426 9.738s-6.197 2.656-4.722-.885c1.476-3.541 9-11.803 9.148-8.853z"
-        fill="#e0c4b9"
-      />
-      <path
-        d="M55.44 231.63c-3.833.055-10.316 9.923-12.188 14.469-2.066 5.016.284 13.275-1.781 17.406-2.066 4.131-4.468 10.553-6.239 17.34-1.77 6.787-4.085 22.502-5.855 32.535-1.03 5.832-2.214 10.151-2.656 13.281-.01.003-.023-.003-.031 0-1.76.631 1.724 4.875.55 8.399-1.181 3.54-4.427 9.386-5.903 12.043-1.475 2.655-2.958 7.382-3.843 11.219-.886 3.836-5.875 11.189-4.99 12.369.885 1.18 2.71-3.287 4.185-7.124s2.666-6.598 2.371-3.942c-.295 2.656-.86 12.716-.266 10.061.61-2.728 1.3-16.55 4.549-19.362 4.914-4.87 6.882-11.726 6.128-18.132.384-.719 1.624-2.102 2.029-3.587.885-3.246 13.784-33.315 15.021-38.489 1.597-6.682 1.176-7.706 2.356-13.017 1.18-5.311 4.112-13.272 7.063-19.469.96-4.06 4.913-24.453-.5-26z"
-        fill="#7e4c31"
-      />
-      <path
-        d="M38.24 267.65c-2.066 5.016-2.95 10.918-4.721 15.64-1.77 4.721-1.77 15.344-2.951 20.066-1.18 4.721-4.427 15.344-4.722 21.246-.295 5.902-2.341 5.646-6.436 10.811-4.096 5.165-2.71 7.309-7.217 16.161-1.18 2.36-9.082 15.185-10.852 14.89-1.77-.295-.349-1.368.133-2.616 1.094-2.834 8.143-10.033 9.028-11.803.886-1.77-4.426 1.77-5.901 3.836s-4.013 3.458-2.537 1.393c1.475-2.066 7.211-7.802 8.687-10.752 1.475-2.951 1.227-6.28.342-8.345-.886-2.066 2.085.098 3.556-.601 1.47-.7 3.526-2.055 4.706-4.12s7.967-25.969 8.853-28.624c.885-2.656 2.95-20.066 2.95-22.426 0-2.36 3.611-9.829 4.496-11.6.886-1.77 2.291-5.81 2.291-7.876s-.295-9.147-.295-9.147 3.558-16.262 3.558-14.196c0 2.065-1.787 26.295-2.968 28.065z"
-        fill="#ddbdb0"
-      />
-      <path
-        d="M47.42 225.16c1.18-7.082.557-18 5.278-20.066 4.722-2.066 11.213 1.18 10.623 9.443-.59 8.262.886 12.099-2.065 12.689-5.223-7.822-14.366 5.809-13.836-2.066z"
-        fill="#7e4c31"
-      />
-      <path
-        fill="#d9ae78"
-        d="M112.17 370.85c.993.744 8.437 3.97 8.437 3.97s21.588-24.814 23.325-27.047c1.737-2.233 8.188-12.655 7.196-12.903-.993-.248-3.97-1.24-4.963-1.985-.992-.744-30.273 29.777-32.258 33.995-1.985 4.218-1.985 4.715-1.737 3.97z"
-      />
-      <path
-        d="M61.3 318.74c1.489 5.21 3.722 13.151 9.43 15.881 5.707 2.73 8.436 3.226 11.91 6.204 3.474 2.977 27.25 19.685 26.055 24.566-.178.728-33.076 1.152-35.98 5.459-2.122 3.147-11.663-50.87-11.414-52.11zM110.18 278.79c-1.737 6.948-14.888 21.836-19.355 27.544-4.466 5.707-8.933 20.347-8.933 20.347s2.978-13.648 13.151-22.333c10.174-8.685 14.64-18.362 15.385-21.588.744-3.226-2.73 14.64-9.678 19.355-6.947 4.715-12.406 20.596-12.654 22.084-.249 1.489 1.985-9.43 11.165-20.099 9.182-10.67 16.875-10.918 16.875-10.918s-6.452 10.174-9.678 10.174-9.677 6.7-11.414 9.181c-1.737 2.482-3.226 3.474 3.97-1.737 7.196-5.21 16.377-7.94 17.866-11.663 1.49-3.722-5.459 16.874-12.159 20.844-6.7 3.97-9.925 9.678-9.925 9.678s19.85-10.67 24.318-26.303c4.466-15.633 3.225-9.43 3.474-11.911.248-2.482 3.97-.745 2.977-3.97-.992-3.226-6.7-7.445-9.18-11.415-2.482-3.97-3.475-4.714-7.197-5.459-3.722-.744-6.203 2.234-15.137 4.963-8.933 2.73-13.15 6.7-15.633 9.181-2.481 2.482-.248-.248 4.467-2.73s14.64-6.947 21.092-6.699c6.451.248 5.707 3.226 6.203 3.474zM133.26 309.56c-.248 3.226-2.233 3.97-3.722 5.707s2.73 3.97-2.482 8.685-2.48 5.955-8.933 4.715-14.144 4.963-15.385 5.955c-1.24.993 3.226-2.481 6.7-3.722s9.181 0 12.903-1.489c3.722-1.489 7.94-9.677 6.7-7.692-1.24 1.985-8.685 6.7-15.88 11.911-7.197 5.21-4.22 7.692-4.22 7.692s-3.721-.496 4.22-6.451c7.94-5.956 11.91-2.234 15.136-7.693 3.226-5.459-9.181 5.211-13.15 11.414-3.971 6.204-2.234 7.196-2.234 7.196s-.745-5.707 7.196-11.414c7.94-5.707 11.663-9.429 11.663-9.429s-14.64 13.896-16.377 18.114c-1.737 4.219-1.985 4.219 2.481-.248s10.67-11.91 12.903-14.392c2.234-2.481 1.49.993-2.481 7.196-3.97 6.204-3.722 3.474 1.489-.496 5.21-3.97 14.144-2.73 14.144-5.956 0-3.225-.497-11.663-3.474-15.633-2.978-3.97-7.196-3.722-7.196-3.97z"
-        fill="#ccc"
-      />
-      <path
-        fill="#ccc"
-        d="M56.34 300.38c3.474 9.43 5.955 11.414 8.189 9.181 2.233-2.233 1.488 1.737.248 2.978s-.993-1.24 4.218.496c5.211 1.737 11.414-3.722 8.189.248-3.226 3.97-11.414 2.978-12.655 5.956-1.241 2.977-1.985 9.18-1.737 11.166.248 1.985-2.73-12.655-4.219-15.633-1.488-2.978-3.474-14.144-3.474-14.144s1.737-2.482 1.241-.248z"
-      />
-      <path
-        fill="#1a1a1a"
-        d="M108.2 367.13c.993-4.715 6.948-12.407 10.918-16.377 3.97-3.97 18.61-15.385 20.844-18.859 2.233-3.474 4.218-3.722 4.218-3.722l2.73 5.459s-14.64 13.151-19.107 19.355c-4.467 6.204-13.896 17.37-14.144 18.859-.248 1.489-5.211-4.467-5.46-4.715z"
-      />
-      <path
-        d="M71.76 371.12c1.11-4.728 33.645-6.362 35.699-6.474 1.998-.109 5.46 6.7 5.46 6.7s-8.438 1.489-14.889 2.481c-6.452.993-18.859 3.474-21.34 5.46-2.481 1.984-2.73 1.24-2.73 1.24l-2.2-9.407z"
-        fill="#1a1a1a"
-      />
-      <path
-        d="M138.21 354c.974 10.506 14.036 14.473 16.546 10.59 1.214-1.878-11.022-12.698-12.68-13.883-1.659-1.185-3.865 4.24-3.865 3.293zM188.83 392.96c.83 2.25 13.391 14.724 15.05 14.961 1.659.237 2.012.751 1.538-1.381-.474-2.133-11.84-14.635-13.262-15.583-1.422-.947-4.156-.247-3.327 2.003z"
-        fill="#a67054"
-      />
-      <path d="M139.74 387.03c.85-3.398 7.009-15.292 8.708-18.053 1.7-2.761 6.16-7.859 6.16-7.859s18.266 8.496 21.664 11.257c3.398 2.762 16.779 17.416 16.779 18.478s-5.098 8.709-7.859 11.47c-2.761 2.76-5.31 8.92-6.584 8.07-1.275-.85-16.567-15.505-20.39-17.629-2.173-1.207-18.903-6.159-18.478-5.734zM188.8 420.16c2.549-2.974 9.345-9.558 11.044-11.257 1.7-1.7 4.673-3.61 6.16-2.761 1.486.85 15.504 17.204 14.867 18.69-.637 1.488-12.106 14.869-13.806 15.08-1.699.213-17.204-17.84-18.266-19.752z" />
-      <path
-        fill="#371a12"
-        d="M200.55 501.04c-1.574-2.923.45-9.893-.225-13.265-.674-3.372-6.07-9.218-6.97-13.265-.899-4.047-2.473-12.815-6.295-16.188-3.822-3.372-2.698-4.047-.899-3.372 1.799.674 7.195-8.094 8.768-7.645 1.574.45.9 2.923 3.148 5.846s3.597-4.272 4.946-.674 11.466 11.466 7.644 17.762c-3.822 6.295-6.52 21.809-8.093 27.204-1.574 5.395-1.574 4.046-2.024 3.597z"
-      />
-      <path
-        fill="#d9ae78"
-        d="M202.12 502.16c1.124-2.248 12.815-28.553 13.49-32.151.674-3.597 4.721-10.342 4.497-11.242-.225-.9-4.272 5.396-5.846 8.768-1.574 3.373-10.342 26.755-11.466 29.453-1.124 2.698-2.473 4.947-.674 5.171z"
-      />
-      <path
-        d="M201.15 503.19c4.995-1.313 13.956-31.033 14.918-32.953.95-1.896 3.822-3.597 4.272-2.024.45 1.574-11.017 35.523-13.04 35.973-2.024.45-5.925-.096-6.15-.996z"
-        fill="#ccc"
-      />
-      <path
-        fill="#d9ae78"
-        d="M220.81 424.94c1.496 1.829 4.489 3.658 6.65 3.658 2.162 0 14.298.498 16.626 1.828 2.328 1.33 9.976 5.32 9.976 5.32s-9.81 24.274-12.137 28.763c-2.328 4.488-6.983 13.966-9.144 16.459-2.162 2.493-11.472 19.12-13.8 19.12-2.327 0-4.82-2.494-5.818-3.824-.998-1.33 1.163-8.313 2.327-11.97 1.164-3.658 3.325-19.785.166-23.442-3.158-3.658-5.652-5.487-7.315-10.474-1.662-4.988-3.159-7.648-3.99-9.144-.831-1.497 2.66-.832 3.491-1.497.832-.665 12.636-14.963 12.968-14.797z"
-      />
-      <path d="M211.17 437.57c2.993 3.325 6.816 2.494 6.484 5.819-.333 3.325 3.159 7.98 4.323 12.636 1.163 4.655 5.652 9.643 2.493 17.789-3.158 8.147-5.036 17.048-7.364 20.041-2.327 2.993-3.11 3.401-3.775 1.572-.665-1.828 5.487-27.93 3.99-30.757-1.496-2.826-7.481-7.814-8.312-10.807-.832-2.992-6.857-11.735-6.026-12.733.832-.997 3.698-.235 4.53-.733.83-.5 3.49-2.827 3.657-2.827z" />
-      <path
-        d="M219.15 441.56c3.824-5.487 5.819-12.137 12.636-10.64 6.816 1.496 4.323 2.494 4.655 8.978.333 6.484 4.969 6.797 2.808 11.95-2.162 5.155-7.316 21.282-12.47 26.768-5.153 5.486-10.122 15.813-9.623 13.818.498-1.995 7.969-23.756 7.138-25.585-.832-1.829-5.143-17.31-5.974-19.97-.832-2.66-.832-3.658.83-5.32z"
-        fill="#371a12"
-      />
-      <path
-        d="M247.25 433.25c-2.993 4.323-5.974 16.176-7.304 21.164-1.33 4.988-4.5 12.586-7.493 17.408-2.992 4.821-7.784 12.284-10.444 17.105-2.66 4.821-6.348 9.828-7.179 8.997-.831-.831 1.164-1.995 2.328-4.988s9.144-13.3 10.64-16.293c1.496-2.992 10.142-21.28 10.807-24.772.665-3.491.166-8.146 2.826-13.134 2.66-4.988 5.653-6.816 5.653-6.816l.166 1.33z"
-        fill="#d9ae78"
-      />
-      <path
-        d="M215.83 497.92c2.328-4.988 15.98-24.821 16.977-27.481.998-2.66 5.966-16.078 7.795-21.232s5.936-16.44 6.601-16.939c.665-.499 7.111 2.21 7.18 3.639.052 1.114-7.248 17.32-8.412 21.145-1.163 3.823-13.32 25.789-15.48 29.28-2.162 3.491-9.337 13.51-10.974 13.848-2.767.57-3.937-1.373-3.687-2.26z"
-        fill="#ccc"
-      />
-      <path
-        fill="#fff"
-        d="M211.08 453.34c1.017-1.829 6.737-4.459 7.901-4.791 1.164-.333-.997 2.493-3.491 4.322-2.494 1.83 1.33-.498 2.992-1.496 1.663-.998 2.827-.998 2.328.499-.499 1.496-4.156 2.494-5.985 4.322-1.829 1.83 2.327-1.496 4.821-2.493 2.494-.998 4.01-1.194 1.35.468s-4.656 6.318-7.15 5.32c-2.493-.997-3.782-4.322-2.766-6.15z"
-      />
-      <path d="M139.85 387.05c-2.698 3.148-10.792 11.916-11.242 13.04-.45 1.124 34.849 27.879 36.198 28.329 5.648-2.233 13.11-11.893 12.985-15.779l-32.787-30.366-5.154 4.777zM185.18 421.47c-2.473 1.574-11.162 15.264-12.286 16.164-1.124.9 15.288 17.087 15.288 17.087s13.498-13.609 13.618-14.607c.266-2.217-15.346-19.469-16.62-18.644z" />
-      <path
-        fill="#fff"
-        d="M190.43 463.72c3.822-3.148 6.008-5.188 4.434-3.39s-1.511 4.065-2.635 4.514c-1.125.45 6.007-5.638 5.108-3.84-.9 1.8-7.132 7.662-5.333 6.089 1.798-1.574 8.256-7.437 5.558-2.94-2.698 4.496-4.659 5.413-5.783 4.738-1.124-.674-1.124-5.171-1.349-5.171z"
-      />
-    </svg>
-  );
-}
-
-export default SvgSlamdunk;
