--- src/iconComponents/index.tsx
@@ -1,2 +1,0 @@
-export { default as OfcWork } from './OfcWork';
-export { default as Slamdunk } from './Slamdunk';
