+++ src/icons/index.ts
@@ -1,0 +1,2 @@
+export {default as ofcWorkSvg} from './ofcWork.svg';
+export {default as slamDunkSvg} from './slamdunk.svg';
